% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aerial_image}
\alias{aerial_image}
\title{A 1 m resolution aerial image.}
\format{
An array with 1000 rows, 1000 columns and 4 layers:
\describe{
\item{1}{blue band reflectance values (430 to 490 nm), in the range 0 to 255}
\item{2}{green band reflectance values (535 to 585 nm), in the range 0 to 255}
\item{3}{red band reflectance values (610 to 660 nm), in the range 0 to 255}
\item{4}{near-infrared band reflectance values (835 to 885 nm), in the range 0 to 255}
}
}
\source{
\url{https://www.bluesky-world.com/}
}
\usage{
aerial_image
}
\description{
A dataset containing image reflectance values for the area bounded by 169000, 170000, 12000,
13000  (xmin, xmax, ymin, ymax) using the Ordance Survey GB Grid Reference system (CRS: +init=epsg:27700).
}
\keyword{datasets}
