% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solartools.R
\name{airmasscoef}
\alias{airmasscoef}
\title{Calculates the airmass coefficient}
\usage{
airmasscoef(localtime, lat, long, julian, merid = round(long/15, 0) * 15,
  dst = 0)
}
\arguments{
\item{localtime}{local time (decimal hour, 24 hour clock).}

\item{lat}{latitude of the location for which the airmass coefficient is required (decimal degrees, -ve south of equator).}

\item{long}{longitude of the location for which the airmass coefficient is required (decimal degrees, -ve west of Greenwich meridian).}

\item{julian}{a numeric value representing the Julian day as returned by \code{\link[=julday]{julday()}}.}

\item{merid}{an optional numeric value representing the longitude (decimal degrees) of the local time zone meridian (0 for GMT). Default is \code{round(long / 15, 0) * 15}}

\item{dst}{an optional numeric value representing the time difference from the timezone meridian (hours, e.g. +1 for BST if \code{merid} = 0).}
}
\value{
the airmass coefficient, i.e. the direct optical path length of a solar beam through the Earth’s atmosphere, expressed as a ratio relative to the path length vertically upwards for a given location and time.
}
\description{
\code{airmasscoef} is used to calculate, for a given location and time, the direct optical path length of a solar beam through the atmosphere of the Earth, expressed as a ratio relative to the path length vertically upwards.
}
\examples{
# airmass coefficient at noon on 21 June 2010, Porthleven, Cornwall
jd <- julday (2010, 6, 21) # Julian day
airmasscoef(12, 50.08, -5.31, jd)
}
