% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiationtools.R
\name{albedo2}
\alias{albedo2}
\title{Partitions surface albedo between ground and canopy albedo}
\usage{
albedo2(alb, fr, ground = TRUE)
}
\arguments{
\item{alb}{a raster object, two-dimensional array or matrix of surface albedo values (range 0 - 1) derived using \code{\link[=albedo]{albedo()}} or \code{\link[=albedo_adjust]{albedo_adjust()}}.}

\item{fr}{a raster object, two-dimensional array or matrix of fractional canopy cover as returned by \code{\link[=canopy]{canopy()}}.}

\item{ground}{a logical value indicating whether to return ground albedo (TRUE) or canopy albedo (FALSE).}
}
\value{
If ground is \code{TRUE}, a raster object or a two-dimensional array of numeric values representing ground albedo (range 0 to 1).

If ground is \code{FALSE}, a raster object or two-dimensional array of numeric values representing canopy albedo (range 0 to 1).
}
\description{
\code{albedo2} is used to calculate either bare ground or canopy albedo from surface albedo.
}
\details{
If alb is a raster object, a raster object is returned. For calculation of net radiation, both
ground and canopy albedo may be are needed. Areas with high canopy cover typically have lower
albedo values than areas with low canopy cover and mean values for these can be derived from
parts of the image with very high or very low canopy cover. It is assumed that albedo of the
image as returned by \code{\link[=albedo]{albedo()}} is a function of both, weighted by canopy cover, such that
canopy albedos are closer to the image-derived albedos in areas of high canopy cover and ground
albedos closer to the image-derived albedo in areas with low canopy cover.
}
\examples{
library(raster)
# ======================
# Calculate image albedo
# ======================
alb <- albedo(aerial_image[,,1], aerial_image[,,2], aerial_image[,,3],
            aerial_image[,,4])
# ======================
# Calculate canopy cover
# ======================
l <- lai(aerial_image[,,3], aerial_image[,,4])
fr <- canopy(l)
# ===========================================
# Calculate and plot ground and canopy albedo
# ===========================================
ag <- albedo2(alb, fr)
ac <- albedo2(alb, fr, ground = FALSE)
par(mfrow=c(2, 1))
plot(if_raster(ag, dtm1m), main = "Ground albedo", col = gray(0:255/255))
plot(if_raster(ac, dtm1m), main = "Canopy albedo", col = gray(0:255/255))
}
