% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiationtools.R
\name{albedo_adjust}
\alias{albedo_adjust}
\title{Adjusts albedo to correct for image brightness and contrast}
\usage{
albedo_adjust(alb_image, alb_modis)
}
\arguments{
\item{alb_image}{a SpatRaster object with numeric albedo values derived from high-resolution aerial imagery, as derived by \code{\link[=albedo]{albedo()}} and converted to a SpatRaster object.}

\item{alb_modis}{a SpatRaster object with numeric albedo values derived from MODIS imagery covering the same extent as \code{alb.image} (range 0 to 1). If the extent of the \code{alb.modis} is greater than that of \code{alb.image}, \code{alb.modis} is cropped.}
}
\value{
a SpatRaster object with numeric values representing the adjusted surface albedo values (range 0 to 1).
}
\description{
\code{albedo_adjust} is used to correct albedo derived from aerial imagery for image brightness and contrast using MODIS data.
}
\examples{
library(terra)
alb <- albedo(aerial_image[,,1], aerial_image[,,2], aerial_image[,,3],
              aerial_image[,,4])
img <- if_raster(alb, dtm1m)
mds <- rast(modis)
ext(mds)<-c(169000, 170000, 12000, 13000)
alb2 <- albedo_adjust(img, mds)
par(mfrow=c(2, 1))
plot(img, main = "Raw albedo", col = gray(0:255/255))
plot(alb2, main = "Adjusted albedo", col = gray(0:255/255))
}
