% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othertools.R
\name{arrayspline}
\alias{arrayspline}
\title{Applies a spline function to an array of values}
\usage{
arrayspline(a, tme, nfact = 24, out = NA)
}
\arguments{
\item{a}{a three-dimensional array (row, column, time)}

\item{tme}{an object of class POSIXct of times for \code{a}. I.e. \code{length(tme) = dim(a)[3]}}

\item{nfact}{indicates the time interval for which outputs are required. E.g to derive hourly from daily data \code{nfact = 24}, or derive six-hourly from daily data \code{nfact = 4}}

\item{out}{an optional character vector indicating the time for which the output is required. Format must be as for \code{tme}.}
}
\value{
If out is unspecified, a three dimensional array of size \code{c(a[1], a[2], (length(tme) - 1) * nfact + 1)} is returned.
If out is specified, a three dimensional array of size \code{c(a[1], a[2], length(out))} is returned.
}
\description{
\code{arrayspline} is used to derive e.g. hourly climate data from e.g. daily values.
}
\details{
\code{arrayspline} uses the Forsythe, Malcolm and Moler method of splining, as specified by
\code{"fmm"} in \code{\link[=spline]{spline()}}. If \code{a[i, j, ]} is a vector of NAs, then tthe corresponding vector in
the output array is also a vector of NAs. It is assumed that all spatial data within \code{a} have
equivelent times. I.e. the time of both \code{a[1, 1, 1]} and \code{a[1000, 1000, 1]} is identical and
equal to \code{tme[1]}.
}
\examples{
library(terra)
tme <- as.POSIXct(c(0:364) * 24 * 3600, origin="2010-01-01", tz = "GMT")
h <- arrayspline(huss, tme, out = "2010-05-01 11:00")
plot(rast(h),
     main = "Specific humidity 2010-05-01 11:00")
}
