% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadtools.R
\name{basinmerge}
\alias{basinmerge}
\title{Merges adjoining basins}
\usage{
basinmerge(dem, basins, boundary)
}
\arguments{
\item{dem}{a raster object, two-dimensional array or matrix of elevations.}

\item{basins}{a raster object, two-dimensional array or matrix with basins numbered as integers as returned by \code{\link[=basindelin]{basindelin()}}.}

\item{boundary}{a single numeric value. Basins seperated by boundaries below this height are merged (should have same units as \code{dtm}).}
}
\value{
a raster object, two-dimensional array or matrix with basins numbered as integers.
}
\description{
\code{basinmerge} merges adjoining basins if the height differences between the
bottom of the basin and the pour point is less than than that
specified by \code{boundary}.
}
\details{
If \code{dem} is a raster object, then a raster object is returned.
If the differences in height between the pour-point and bottom of the basin is
less than that specified by \code{boundary} the basin is merged with basin to which
water or air would pour.
}
\examples{
basins2 <- basinmerge(dtm100m, basins100m, 1)
par(mfrow=c(1, 2))
plot(basins100m, main = "Basins")
plot(basins2, main = "Merged basins")
}
