% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basins100m}
\alias{basins100m}
\title{A raster of basins numbered as integers}
\format{A raster object with 187 rows and 214 columns.}
\usage{
basins100m
}
\description{
A raster object of basins numbered as integers for the area bounded by 160000, 181400,
11300, 30000  (xmin, xmax, ymin, ymax) using the Ordance Survey GB Grid Reference system
(CRS: +init=epsg:27700) as returned by \code{\link[=basindelin]{basindelin()}}.
}
\keyword{datasets}
