% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadtools.R
\name{basinsort}
\alias{basinsort}
\title{Orders drainage basins by elevation}
\usage{
basinsort(dem, basins)
}
\arguments{
\item{dem}{a raster object, two-dimensional array or matrix of elevations.}

\item{basins}{a raster object, two-dimensional array or matrix of basins numbered as integers.}
}
\value{
a raster object, two-dimensional array or matrix of sequentially numbered basins
}
\description{
\code{basinsort} is an internal function used by \code{\link[=basindelin]{basindelin()}} and \code{\link[=basinmerge]{basinmerge()}} to number
drainage basins sequentially from lowest elevation (of lowest point) to highest.
}
\examples{
basins <- matrix(c(1:4), nrow = 2, ncol = 2)
dem <- matrix(c(4:1), nrow = 2, ncol = 2)
basinsort(dem, basins)
}
\keyword{internal}
