% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadtools.R
\name{cadconditions}
\alias{cadconditions}
\title{Calculates whether conditions are right for cold air drainage}
\usage{
cadconditions(h, tc, n, p = 100346.13, wind, startjul, lat, long,
  starttime = 0, hourint = 1, windthresh = 4.5, emthresh = 0.5,
  tz = 0, dst = 0, con = TRUE)
}
\arguments{
\item{h}{a vector of hourly specific humidities (\ifelse{html}{\out{kg kg<sup>-1</sup> }}{\eqn{kg kg^{-1}}}).}

\item{tc}{a single numeric value, raster object, two-dimensional array or matrix of temperatures (ºC).}

\item{n}{a vector of hourly fractional cloud cover values (range 0 - 1).}

\item{p}{an optional vector of hourly atmospheric pressure values (Pa).}

\item{wind}{a vector of wind speed values at one metre above the ground (\ifelse{html}{\out{m s<sup>-1</sup> }}{\eqn{m s^{-1}}})}

\item{startjul}{julian day of first observation as returned by \code{\link[=julday]{julday()}}}

\item{lat}{latitude of the location for which cold air drainage conditions are required (decimal degrees, -ve south of equator).}

\item{long}{longitude of the location for which cold air drainage conditions are required (decimal degrees, -ve west of Greenwich meridian).}

\item{starttime}{the hour of the first observation (decimal, 0-23).}

\item{hourint}{the interval (in hours) between successive observations}

\item{windthresh}{an optional threshold value of wind speed below which cold air conditions can occur (\ifelse{html}{\out{m s<sup>-1</sup> }}{\eqn{m s^{-1}}})}

\item{emthresh}{an optional threshold value of emissivity below which cold air conditions can occur (range 0 - 1)}

\item{tz}{an optional numeric value specifying the time zones expressed as hours different from GMT (-ve to west).}

\item{dst}{an optional numeric value representing the local summer time adjustment (hours, e.g. +1 for BST).}

\item{con}{an optional logical value indicating whether or not to allow cold air drainage conditions to occur only if conditions are right for three or more consecutive hours. Ignored if hourint != 1.}
}
\value{
a vector of binary values indicating whether cold air drainage conditions occur (1) or not (0)
}
\description{
\code{cadconditions} determines whether wind speed, humidity and cloud cover  are such that
cold air drainage is likely to occur
}
\details{
\code{cadconditions} uses a time series of wind and emissivity data to determine
whether cold air drainage conditions are likely to occur and returns a binary
vector of the same length as \code{em} and \code{wind} indicating whether conditions
occur (1) or not (0). They are assumed to occur at night or within three
hours of dawn only and when both \code{em} and \code{wind} are below the values
specified by \code{windthresh} and \code{emthresh}. If no start time is specified it
is assumed that first index of \code{em} and \code{wind} occurs at midnight on the
date specified by \code{startjul}. If no \code{hourint} is provided, the time
interval between indices of \code{em} and \code{wind} are assumed to be hourly.
If \code{con} is TRUE  and \code{hourint} is 1 (the default), cold air drainage conditions are
assumed to persist only when conditions are right for three consecutive hours or more.
The first index of the output, for which prior conditions cannot be assessed
is set to 1 if conditions are right, irrespective of prior conditions. The
second index is set to one only if conditions are right in both that hour
and the preceding hour. If \code{con} is FALSE or \code{hourint != 1} prior conditions
are ignored.
}
\examples{
# ===============================================
# Mean daily climate for Lizard, Cornwall in 2010
# ===============================================
h <- apply(huss, 3, mean, na.rm = TRUE)
p <- apply(pres, 3, mean, na.rm = TRUE)
tmin <- apply((tas - dtr), 3, mean, na.rm = TRUE)[2:364]
tmax <- apply((tas + dtr), 3, mean, na.rm = TRUE)[2:364]
# =====================================
# hourly climate 2nd Jan to 30 Dec 2010
# =====================================
h <- spline(h, n = 8737)$y[13:8724]
p <- spline(p, n = 8737)$y[13:8724]
n <- apply(cfc[,,13:8724], 3, mean)
rdni <- apply(dnirad[,,13:8724], 3, mean)
rdif <- apply(difrad[,,13:8724], 3, mean)
jd <- julday(2010, 2, 1)
jd <- c(jd:(jd+362))
tc <- hourlytemp(jd, h, n, p, rdni, rdif, tmin, tmax, 50.05, -5.19)
ws10m <- spline(wind2010$wind10m, n = 8755)$y[25:8736]
ws1m <- windheight(ws10m, 10, 1)
# =================================================================
# Calculate whether cold air drainage, persists and plot proportion
# =================================================================
startjul <- julday(2010,1,2)
hist(cadconditions(h, tc, n, p, ws1m, startjul, 50.05, -5.19),
     main = "", xlab = "Cold air drainage conditions (1 = Y)")
}
