% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiationtools.R
\name{canopy}
\alias{canopy}
\title{Calculates canopy cover}
\usage{
canopy(l, x)
}
\arguments{
\item{l}{a raster object, two-dimensional array or matrix of leaf area index values as returned by \code{\link[=lai]{lai()}}.}

\item{x}{a raster object, two-dimensional array of numeric values representing the ratio of vertical to horizontal projections of leaf foliage as returned by \code{\link[=leaf_geometry]{leaf_geometry()}}.}
}
\value{
a raster object or a two-dimensional array of numeric values representing fractional canopy cover estimated as the proportion of isotropic radiation transmitted through the canopy.
}
\description{
\code{canopy} is used to calculate fractional canopy cover.
}
\details{
Canopy cover calculated by this function is defined as 1 - the proportion of isotropic
radiation transmitted through the canopy. If \code{l} is a raster object, a raster object is
returned.
}
\examples{
l <- lai(aerial_image[,,3], aerial_image[,,4])
l <- if.raster(l, dtm1m) # convert to raster
x <- leaf_geometry(veg_hgt)
fr <- canopy(l, x)
plot(fr, main = "Fractional canopy cover")
}
