% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cfc}
\alias{cfc}
\title{A 0.05º resolution dataset of hourly fractional cloud cover}
\format{
An array with 5 rows, 8 columns and 8670 hourly values
}
\source{
\url{http://www.cmsaf.eu/}
}
\usage{
cfc
}
\description{
A dataset containing hourly fractional cloud cover values in 2010 for the area bounded by
-5.40, -5.00, 49.90, 50.15 (xmin, xmax, ymin, ymax) (CRS: +init=epsg:4326).
}
\keyword{datasets}
