% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiationtools.R
\name{clearskyrad}
\alias{clearskyrad}
\title{calculates clearsky radiation}
\usage{
clearskyrad(tme, lat, long, h, tc, p = 101300, G = 2.78,
  Ie = 1352.778)
}
\arguments{
\item{tme}{a single value or vector of POSIXlt objects indicating the time(s)
for which clearksy radiation is required.}

\item{lat}{latitude in decimal degrees}

\item{long}{longitude in decimal degrees}

\item{h}{a single value or vector of specific humidities (\ifelse{html}{\out{kg kg<sup>{-1}</sup> }}{\eqn{kg kg^{-1}}}).}

\item{tc}{a single value or vector of temperatures (ºC).}

\item{p}{an optional single value or vector of pressures (Pa).}

\item{G}{an optional single value or vector describing he moisture profile
in the atmosphere (per Smith 1966).}

\item{Ie}{an optional single value for extra-terrestrail radiation to permit adjustment for
sun-earth distances (see details).}
}
\value{
a single value or vector of clearksy radiation.
}
\description{
\code{clearskyrad} is used to calculate clear-sky shortwave irradiance using the
Crawford & Duchon (1999) method
}
\details{
The units returned are the same as for \code{Ie}, with the default option in W / M^2.
If no values for \code{p} are provided, a default value of 101300 Pa, typical of
sea-level pressure, is assumed. The method used is that detailed in
Crawford & Duchon (1999) Quarterly Journal of the Royal Meteorological
Society 122: 1127-1151.
}
\examples{
tme <- as.POSIXlt(c(0:23) * 3600, origin = "2010-05-23 00:00", tz = "GMT")
Io <- clearskyrad(tme, 50, -5, 0.007953766 , 11)
plot(Io ~ as.POSIXct(tme), type = "l")
}
\seealso{
The function \code{\link[=cloudfromrad]{cloudfromrad()}} uses this function to return 1 - ratio of measured to clearsky
radiation for input when computing longwave radiation when using \code{\link[=longwavetopo]{longwavetopo()}} or \code{\link[=longwaveveg]{longwaveveg()}}
}
