% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiationtools.R
\name{cloudfromrad}
\alias{cloudfromrad}
\title{calculates cloud cover form shortwave radiation}
\usage{
cloudfromrad(rad, tme, lat, long, h, tc, p = 101300, G = 2.78,
  Ie = 1352.778)
}
\arguments{
\item{rad}{a single numeric value or vector of solar irradiance(s). Units must be the same
as those for \code{Ie}. Default is Watts /m^2.}

\item{tme}{a single value or vector of POSIXlt objects indicating the time(s)
for which clearksy radiation is required.}

\item{lat}{latitude in decimal degrees}

\item{long}{longitude in decimal degrees}

\item{h}{a single value or vector of specific humidities (\ifelse{html}{\out{kg kg<sup>{-1}</sup> }}{\eqn{kg kg^{-1}}}).}

\item{tc}{a single value or vector of temperatures (ºC).}

\item{p}{an optional single value or vector of pressures (Pa).}

\item{G}{an optional single value or vector describing he moisture profile
in the atmosphere (per Smith 1966).}

\item{Ie}{an optional single value for extra-terrestrail radiation to permit adjustment for
sun-earth distances (see details).}
}
\value{
a single value or vector of clearksy radiation.
}
\description{
\code{cloudfromrad} is used to derive a cloud cover index as 1 - the ratio of measured to clearksy radiation
missing values (e.g. at night) are interpolated
}
\details{
If no values for \code{p} are provided, a default value of 101300 Pa, typical of
sea-level pressure, is assumed. The method used is that detailed in
Crawford & Duchon (1999) Quarterly Journal of the Royal Meteorological
Society 122: 1127-1151.
}
\examples{
tme <- as.POSIXlt(c(0:23) * 3600, origin = "2010-05-23 00:00", tz = "GMT")
rad <- clearskyrad(tme, 50, -5, 0.007953766 , 11) * 0.75
cfc <- cloudfromrad(rad, tme, 50, -5, 0.007953766 , 11)
plot(cfc ~ as.POSIXct(tme), type = "l") # should be 0.25
}
\seealso{
The function \code{\link[=clearskyrad]{clearskyrad()}} is uses to derive clear-sky irradiance. Can be used
to derive cloud covers for computing longwave radiation when using \code{\link[=longwavetopo]{longwavetopo()}} or \code{\link[=longwaveveg]{longwaveveg()}}
}
