% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools.R
\name{coastalNCEP}
\alias{coastalNCEP}
\title{Calculates coastal effects from NCEP data}
\usage{
coastalNCEP(
  landsea,
  ncephourly,
  steps = 8,
  use.raster = T,
  zmin = 0,
  plot.progress = T,
  tidyr = F
)
}
\arguments{
\item{landsea}{a raster object with NAs (representing sea) or any non-NA value and a projection system defined.}

\item{ncephourly}{a dataframe of hourly climate data as returned by \code{\link[=hourlyNCEP]{hourlyNCEP()}}.}

\item{steps}{an optional integer. Coastal effects are calculated in specified directions upwind. Steps defines the total number of directions used. If the default 8 is specified, coastal effects are calculated at 45º intervals.}

\item{use.raster}{an optional logical value indicating whether to mask the output values by \code{landsea}.}

\item{zmin}{optional assumed sea-level height. Values below this are set to zmin}

\item{plot.progress}{logical value indicating whether to produce plots to track progress.}

\item{tidyr}{logical value indicating whether to download 30m resolution digital elevation data
and use these data to improve the assignment of land and sea pixels in \code{landsea}.}
}
\value{
a three-dimension array of hourly temperature values for each pixel of \code{landsea}
}
\description{
Calculates coastal effects from NCEP data
}
\details{
coastalNCEP downloads digital elevation and varying resolutions to calculate
coastal effects by applying a variant of \code{\link[=invls]{invls()}} over a greater extent then that specified by landsea, but the resulting outputs are
cropped to the same extent as landsea. Land temperatyres as a function of coastal exposure
and sea-surface temperature data are then calculated. Sea surface temperature data are
automatically downloaded from the NOAA.
}
\examples{
library(raster)
# Download NCEP data
ll <- latlongfromraster(dtm100m)
tme <-as.POSIXlt(c(0:31) * 3600 * 24, origin = "2015-03-15", tz = "GMT")
ncephourly<-hourlyNCEP(NA, ll$lat, ll$long, tme)
aout <- coastalNCEP(dtm100m, ncephourly)
# Calculate mean temperature and convert to raster
mtemp <- if_raster(apply(aout, c(1, 2), mean), dtm100m)
plot(mtemp, main = "Mean temperature")
}
