% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demworld}
\alias{demworld}
\title{A 1.875 (long) by  1.904129(lat) resolution dataset of elevations}
\format{
A raster object with 94 rows and 192 columns.
}
\source{
\url{http://www.cmsaf.eu/}
}
\usage{
demworld
}
\description{
A dataset global dataset elevation in metres (CRS: +init=epsg:4326)
}
\keyword{datasets}
