% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{difrad}
\alias{difrad}
\title{A 0.05º resolution dataset of hourly diffuse radiation}
\format{An array with 5 rows, 8 columns and 8670 hourly values}
\source{
\url{http://www.cmsaf.eu/}
}
\usage{
difrad
}
\description{
A dataset containing hourly diffuse radiation values in 2010 (\ifelse{html}{\out{MJ m<sup>-2</sup> hr<sup>-1</sup>}}{\eqn{MJ m^{-2} hr^{-1}}})
for the area bounded by -5.40, -5.00, 49.90, 50.15 (xmin, xmax, ymin, ymax) (CRS: +init=epsg:4326).
}
\keyword{datasets}
