% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadtools.R
\name{.hgttongbr}
\alias{.hgttongbr}
\title{Internal funcion to calculate whether neighbouring cell is higher or lower}
\usage{
.hgttongbr(m)
}
\description{
Internal funcion to calculate whether neighbouring cell is higher or lower
}
