% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools.R
\name{.invls.auto}
\alias{.invls.auto}
\title{Calculates coastal exposure automatically}
\usage{
.invls.auto(r, steps = 8, use.raster = T, zmin = 0,
  plot.progress = T)
}
\description{
Calculates coastal exposure automatically
}
