% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools.R
\name{.micro_ncep2}
\alias{.micro_ncep2}
\title{Temporary NMR Function}
\usage{
.micro_ncep2(loc = c(-5.3, 50.13), dstart = "01/01/2017",
  dfinish = "31/12/2017", dem = NA, dem2 = dem,
  nyears = as.numeric(substr(dfinish, 7, 10)) - as.numeric(substr(dstart,
  7, 10)) + 1, REFL = 0.15, slope = NA, aspect = NA, DEP = c(0,
  2.5, 5, 10, 15, 20, 30, 50, 100, 200), Refhyt = 2, Usrhyt = 0.01,
  Z01 = 0, Z02 = 0, ZH1 = 0, ZH2 = 0, run.gads = 1,
  write_input = 0, writecsv = 0, reanalysis = TRUE, windfac = 1,
  warm = 0, ERR = 1.5, RUF = 0.004, EC = 0.0167238, SLE = 0.95,
  Thcond = 2.5, Density = 2.56, SpecHeat = 870, BulkDensity = 1.3,
  PCTWET = 0, rainwet = 1.5, cap = 1, CMH2O = 1, hori = rep(0,
  24), runmoist = 1, PE = rep(1.1, 19), KS = rep(0.0037, 19),
  BB = rep(4.5, 19), BD = rep(1.3, 19), DD = rep(2.56, 19),
  maxpool = 10000, rainmult = 1, evenrain = 0,
  SoilMoist_Init = c(0.1, 0.12, 0.15, 0.2, 0.25, 0.3, 0.3, 0.3, 0.3,
  0.3), L = c(0, 0, 8.2, 8, 7.8, 7.4, 7.1, 6.4, 5.8, 4.8, 4, 1.8, 0.9,
  0.6, 0.8, 0.4, 0.4, 0, 0) * 10000, R1 = 0.001, RW = 2.5e+10,
  RL = 2e+06, PC = -1500, SP = 10, IM = 1e-06, MAXCOUNT = 500,
  LAI = 0.1, LOR = 1, snowmodel = 1, snowtemp = 1.5,
  snowdens = 0.375, densfun = c(0.5979, 0.2178, 0.001, 0.0038),
  snowmelt = 1, undercatch = 1, rainmelt = 0.0125, shore = 0,
  tides = 0, hourly = 1, rainhourly = 0, rainhour = 0,
  rainoff = 0, lamb = 0, IUV = 0, soilgrids = 0, message = 0,
  fail = nyears * 24 * 365, spatial = NA, save = 0, snowcond = 0,
  intercept = 0/100 * 0.3, grasshade = 0, coastal = T)
}
\description{
Temporary NMR Function
}
