% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools.R
\name{.shortwave.ts}
\alias{.shortwave.ts}
\title{get shortwave radiation (time series)}
\usage{
.shortwave.ts(dni, dif, jd, localtime, lat, long, slope, aspect, ha = 0,
  svv = 1, x = 1, l = 0, albr = 0, merid = round(long/15, 0) *
  15, dst = 0, difani = TRUE)
}
\description{
get shortwave radiation (time series)
}
