% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runauto2.R
\name{.shortwaveveg2}
\alias{.shortwaveveg2}
\title{Downscaled shortwave radiation}
\usage{
.shortwaveveg2(
  dni,
  dif,
  julian,
  localtime,
  lat = NA,
  long = NA,
  dtm = array(0, dim = c(1, 1)),
  slope = NA,
  aspect = NA,
  svv = 1,
  albg = 0.15,
  albc = 0.23,
  fr,
  albr = 0.23,
  ha = 0,
  res = 1,
  merid = NA,
  dst = 0,
  shadow = TRUE,
  x,
  l,
  difani = TRUE
)
}
\description{
Downscaled shortwave radiation
}
