% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dtm1m}
\alias{dtm1m}
\title{A 1 m resolution raster object of elevation for part of the Lizard Peninsula, Cornwall, UK.}
\format{A raster object with 1000 rows and 1000 columns.}
\source{
\url{http://www.tellusgb.ac.uk/}
}
\usage{
dtm1m
}
\description{
A raster object containing elevation in metres with sea coded as NA for the area bounded by
169000, 170000, 12000, 13000  (xmin, xmax, ymin, ymax) using the Ordance
Survey GB Grid Reference system (CRS: +init=epsg:27700).
}
\keyword{datasets}
