% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dtr}
\alias{dtr}
\title{A 1 km resolution dataset of diurnal temperature ranges}
\format{An array with 19 rows, 22 columns and 365 daily values}
\source{
\url{https://www.metoffice.gov.uk/}
}
\usage{
dtr
}
\description{
A spatially interpolated dataset containing diurnal temperature ranges (ºC) in 2010 for the
area bounded by 160000, 182000, 11000, 30000  (xmin, xmax, ymin, ymax) using the Ordance
Survey GB Grid Reference system (CRS: +init=epsg:27700).
}
\keyword{datasets}
