% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othertools.R
\name{fitmicro}
\alias{fitmicro}
\title{Fits micro- or mesoclimate model}
\usage{
fitmicro(microfitdata, alldata = FALSE, windthresh = NA,
  continious = FALSE, iter = 999)
}
\arguments{
\item{microfitdata}{a data.frame with at least the following columns (see, for example, microfit data):
\describe{
\item{temperature}{microclimate temperature readings}
\item{reftemp}{Reference (e.g. coarse-scale or weather station) temperatures}
\item{wind}{Wind speeds}
\item{netrad}{Net radiation values}
}}

\item{alldata}{an optional logical value indicating whether to fit the model using all data (TRUE) or using a randomization procedure (FALSE). See details.}

\item{windthresh}{an optional single numeric value indicating the threshold wind speed above which an alternative linear relationship between net radiation the microclimate temperature anomoly is fitted. See details.}

\item{continious}{an optional logical value indicating whether to treat wind speed as a continious variable.}

\item{iter}{a single integer specifying the iterations to perform during randomization. Ignored if \code{alldata} = TRUE.}
}
\value{
a data,frame with the following columns:
\describe{
\item{Estimate}{parameter estimates and \code{windthresh}}
\item{Std.Dev}{Standard deviation of parameter estimates}
\item{P}{Two-tailed p-value}
}
}
\description{
\code{fitmicro} is used to fit a micro- or mesoclimate model using field temperature readings, and estimates of reference temperature, net radiation
and wind speed at the locations of those readings.
}
\details{
If modelling mesoclimate, it is assumed that altitudinal, coastal and cold-air
drainage effects have already been accounted for in the calculation of \code{reftemp}.
It is therefore assumed that the most important energy fluxes determining near-surface
temperature are those due to the radiation flux and convection that occurs at
the surface-atmosphere boundary. Heat fluxes into the soil and latent heat
exchange are considered to be small and proportional to the net radiation
flux, and the heat capacity of the vegetation is considered to be small so
that, compared to the time-scale of the model, surface temperature rapidly
reach equilibrium. In consequence, the difference between the near-ground
temperature and the ambient temperature is a linear function of \code{netrad}.
The gradient of this linear relationship is a measure of the thermal
coupling of the surface to the atmosphere. If this relationship is applied
to vegetation, assuming the canopy to act like a surface, while air density
and the specific heat of air at constant pressure are constant, the slope
varies as a function of a wind speed factor, such that different slope values
are assumed under high and low wind conditions.  Hence, as a default, \code{fitmicro}
fits a linear model of the form \code{lm((temperature - reftemp) ~ netrad * windfact)}
where windfact is given by \code{ifelse(wind > windthresh, 1, 0)} If \code{continious} is
set to TRUE, then a linear model of the form \code{lm((temperature - reftemp) ~ netrad * log(wind + 1)}
is fitted. If \code{alldata} is FALSE, random subsets of the data are selected and the analyses repeated
\code{iter} times to reduce the effects of of temporal autocorrelation. Parameter
estimates are derived as the median of all runs. If \code{continious} is set to FALSE
and no value is provided for \code{windthresh}, it is derived by iteratively trying out
different values, and selecting that which yields the best fit. The gradient of
the relationship is also dependent on vegetation structure, and in some
circumstances it may therefore be advisable to fit seperate models for each
vegetation type.
}
\examples{
fitmicro(microfitdata)
fitmicro(mesofitdata, alldata = TRUE)
fitmicro(mesofitdata, alldata = TRUE, continious = TRUE)
}
