% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadtools.R
\name{flowacc}
\alias{flowacc}
\title{Calculates accumulated flow}
\usage{
flowacc(dem, basins)
}
\arguments{
\item{dem}{a raster object, two-dimensional array or matrix of elevations.}

\item{basins}{a raster object, two-dimensional array or matrix with basins numbered as integers as returned by \code{\link[=basindelin]{basindelin()}}.}
}
\value{
a raster object, two-dimensional array or matrix of accumulated flow.
}
\description{
\code{flowacc} is used by \code{\link[=pcad]{pcad()}} to calculate accumulated flow to each cold air drainage
basin
}
\details{
Accumulated flow is expressed in terms of number of cells.
}
\examples{
# Merge basins seperated by boundary < 2m
basins <- basinmerge(dtm100m, basins100m, 2)
# Calculate accumulated flow: takes a few minutes to run
fa <- flowacc(dtm100m, basins)
# plot data (expressed as area)
plot(log(fa * 100^2), main = "Log (accumulated flow)")
}
