% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools.R
\name{get_dem}
\alias{get_dem}
\title{Gets raster of elevations}
\usage{
get_dem(r = NA, lat, long, resolution = 30, zmin = 0, xdims = 200, ydims = 200)
}
\arguments{
\item{r}{optional raster object covering the location for which elevations are
required}

\item{lat}{the latitude (in decimal degrees) of the location for which elevations
are required. Not used if \code{r} is a raster.}

\item{long}{the longitude (in decimal degrees) of the location for which elevations
are required. Not used if \code{r} is a raster}

\item{resolution}{the resolution (in metres) of the required dem (see details).}

\item{zmin}{assumed sea-level height. Values below this are set to zmin.}

\item{xdims}{optional numeric value specifying the number of longitudinal grid cells}

\item{ydims}{optional numeric value specifying the number of latitudinal grid cells}
}
\value{
a raster object of elevations in metres.
}
\description{
Several web services provide access to raster elevations. This function
provides access to the Registry of Open Data on AWS to retrieve a raster object of
elevations.
}
\details{
Currently, \code{get_dem} utilises the Amazon Web Services (AWS)
(https://aws.amazon.com/public-datasets/terrain/) to retrieve dems.  The spatial
resolution of the data available vary by location
(see https://mapzen.com/documentation/terrain-tiles/data-sources/ for details
and https://github.com/tilezen/joerd/blob/master/docs/images/footprints-preview.png
for a map showing the resolution of data available for different parts of the globe).
Global coverage is ~30m resolution and if the resolution of data requested is
higher than that available, elevations are derived by interpolation. If \code{r} is
unspecified, then \code{lat} and \code{long} are used to retrieve a xdims x ydims cell raster
centred on \code{lat} and \code{long}. Bathymetry is also returned automatically form AWS,
so \code{zmin} allows the user to specify a sea-level height. Elevations below \code{zmin}
are set at \code{zmin}. For use with \code{microclima} the returned raster has a Mercator
Coordinate Reference System if the latitude specified by \code{lat} is between 80
degrees S and 84 degrees N. In polar regions WGS 84 / NSIDC Sea Ice Polar
Stereographic projections are used. X, y and Z units are all in metres. If \code{r} is
specified, the Coordinate Reference System of \code{r} is used.
}
\examples{
library(raster)
dem50m <- get_dem(dtm100m, resolution = 50)
plot(dem50m) # 50m raster, OSGB projection system

dem5m <- get_dem(lat = 49.97, long = -5.22, resolution = 5)
plot(dem5m) # 5m raster, Mercator projection system
}
