% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools.R
\name{get_NCEP}
\alias{get_NCEP}
\title{Obtains NCEP data required to run microclima}
\usage{
get_NCEP(lat, long, tme, reanalysis2 = TRUE)
}
\arguments{
\item{lat}{the latitude of the location for which data are required}

\item{long}{the longitude of the location for which data are required}

\item{tme}{a POSIXlt object covering the duration for which data are required.
Hourly data are returned irrespective of the time interval of \code{tme}.}

\item{reanalysis2}{Logical. Should data be obtained from the Reanalysis II dataset (default) or
from Reanalysis I (data prior to 1979).}
}
\value{
a dataframe with the following variables: (i) obs_time:times in YYY-MM-DD HH:MM format (GMT),
(ii) Tk: mean temperatures at 2m (K), (iii) Tkmin:  minimum temperatures at 2m (K) (iv) Tkmax:
maximum temperatures at 2m (K), (v) sh: specific humidity at 2m (Kg / Kg),
(vi) pr: surface pressure (Pa), (vii) wu u wind vector at 10m (m/s) (viii) wv: v wind vector
at 10m (m/s) (ix) dlw: downward longwave radiation flux (\ifelse{html}{\out{W m<sup>-2</sup>}}{\eqn{W m^-2}}),
(x) ulw: upward longwave radiation flux (\ifelse{html}{\out{W m<sup>-2</sup>}}{\eqn{W m^-2}}), (xi)
dsw: downward shortwave radiation flux (\ifelse{html}{\out{W m<sup>-2</sup>}}{\eqn{W m^-2}})
}
\description{
Obtains NCEP data required to run microclima
}
\details{
The NCEP-NCAR and NCEP–DOE Atmospheric Model Intercomparison Project (Kanamitso et al 2002)
provides estimates of a range of climate and radiation data for the period 1948 to present.
Data at six-hourly intervals are available at a spatial resolution of ~ 1.8 to 2.5º. The function downloads
the following datasets: mean, min and max air temperature at 2m,  specific humidity and 2m, surface pressure,
the u and v wind vectors at 10, the downward and upward longwave radiation fluxes at the surface
and the downward shortwave flux at the surface. Data for days either side of \code{tme} are downloaded to aid with
interpolation by \code{\link[=hourlyNCEP]{hourlyNCEP()}}. Requires internet connection.
}
\examples{
tme <- as.POSIXlt(c(1:15) * 24 * 3600, origin = "2015-01-15", tz = 'GMT')
head(get_ncep(tme, 50, -5))
}
\references{
Kanamitsu M, Ebisuzaki W, Woollen J, Yang SK, Hnilo JJ, Fiorino M, Potter GL (2002)
Ncep–doe amip-ii reanalysis (r-2). Bulletin of the American Meteorological Society 83: 1631-1644.
}
\seealso{
\code{\link[=hourlyNCEP]{hourlyNCEP()}}
}
