% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{globalclimate}
\alias{globalclimate}
\title{A dataset of global climate variables}
\format{An array with 94 rows, 192 columns and the following five climate vaiables:
\describe{
\item{1}{mean annual temperature (ºC)}
\item{2}{coefficient of variation in temperature (K)}
\item{3}{mean annual temperature (mm per year)}
\item{4}{coefficient of variation in annual rainfall (mm per 0.25 days)}
\item{5}{numeric month with the most rainfall (1-12)}
}}
\source{
\url{http://www.ncep.noaa.gov/}
}
\usage{
globalclimate
}
\description{
A global dataset containing containing the following climate variables averaged
over the period 2008 to 2017 as used by \code{\link[=laifromhabitat]{laifromhabitat()}}.
}
\keyword{datasets}
