% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{habitats}
\alias{habitats}
\title{A table of habitat types}
\format{a data.frame with 17 rows and two columns:
\describe{
\item{number}{Numeric values for each habitat type}
\item{descriptor}{A description of the habitat type}
}
#' @source \url{https://modis.gsfc.nasa.gov/data/}}
\usage{
habitats
}
\description{
A dataset containing habitat descriptors and habitat numbers as used by
\code{\link[=laifromhabitat]{laifromhabitat()}}.
}
\keyword{datasets}
