% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools.R
\name{hourlyNCEP}
\alias{hourlyNCEP}
\title{Obtain NCEP data for running microclima}
\usage{
hourlyNCEP(tme, lat, long, reanalysis2 = TRUE)
}
\arguments{
\item{tme}{a POSIXlt object covering the duration for which data are required.
Hourly data are returned irrespective of the time interval of \code{tme}.}

\item{lat}{the latitude of the location for which data are required}

\item{long}{the longitude of the location for which data are required}

\item{reanalysis2}{Logical. Should data be obtained from the Reanalysis II dataset (default) or
from Reanalysis I (data prior to 1979).}
}
\value{
a dataframe with the following variables: (i) obs_time: a POSIXlt oobject of times,
(ii) temperature: temperatures at 2m (ºC), (iii) humidity: specific humidity at 2m (Kg / Kg),
(iv) pressure: surface pressure (Pa), (v) windspeed: wind speed at 2m (metres per
second), (vi) winddir: wind direction (degrees from N), (vii) emissivity: emissivity of the
atmosphere (0 - 1), (viii) netlong: Net longwave radiation (\ifelse{html}{\out{MJ m<sup>-2</sup> hr<sup>-1</sup>}}{\eqn{MJ m^-2 hr^{-1}}})
(ix) uplong: upward longwave radiation (\ifelse{html}{\out{MJ m<sup>-2</sup> hr<sup>-1</sup>}}{\eqn{MJ m^-2 hr^{-1}}}),
(x) downlong: downward longwave radiation (\ifelse{html}{\out{MJ m<sup>-2</sup> hr<sup>-1</sup>}}{\eqn{MJ m^-2 hr^{-1}}}),
(xi) rad_dni: Direct radiation normal to the solar beam (\ifelse{html}{\out{MJ m<sup>-2</sup> hr<sup>-1</sup>}}{\eqn{MJ m^-2 hr^{-1}}}),
(xii) rad_dif: Diffuse radiation (\ifelse{html}{\out{MJ m<sup>-2</sup> hr<sup>-1</sup>}}{\eqn{MJ m^-2 hr^{-1}}}),
(xii) szenith: the zenith angle (degrees).
}
\description{
\code{hourlyNCEP} downloads the required NCEP climate and radiation forcing data
required for running microclima and interpolates 4x daily data to hourly.
}
\details{
The NCEP-NCAR and NCEP–DOE Atmospheric Model Intercomparison Project (Kanamitso et al 2002)
provides estimates of a range of climate and radiation data for the period 1948 to present.
Data at six-hourly intervals are available at a spatial resolution of 2.5ºC. The function downloads
the following datasets: air temperature at 2m,  specific humidity and 2m, surface pressure,
the u and v wind vectors at 10, thr downward and upward longwave radiation fluxes at the surface
and the downward shortwave flux at the surface. Pressure, humidity and the u and v wind vectors
are converted to hourly using spline interpolation. Wind speeed and diretcion and then
calculated and adjusted to give values at 1 m using \code{\link[=windheight]{windheight()}}. The diffuse radiation
proportion is calculated using \code{\link[=difprop]{difprop()}}, and hourly obtained by adusting for the direction
of the solar beam and airmass thickness using functions \code{\link[=siflat]{siflat()}} and \code{\link[=airmasscoef]{airmasscoef()}}.
Hourly temperature are derived using \code{\link[=hourlytemp]{hourlytemp()}} and longwave radiation by splining emissivity
values. Requires internet connection.
}
\examples{
tme <- as.POSIXlt(c(0:364) * 24 * 3600, origin ="2015-01-01 00:00", tz = "GMT")
# NB takes a while to download data
hdata<- hourlyNCEP(tme, 50, -5)
head(hdata)
plot(temperature ~ as.POSIXct(obs_time), data = hdata, type = "l", xlab = "Month")
plot(rad_dif ~ as.POSIXct(obs_time), data = hdata, type = "l", xlab = "Month")
}
