% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othertools.R
\name{hourlytemp}
\alias{hourlytemp}
\title{Derives hourly temperatures from daily data}
\usage{
hourlytemp(julian, em = NA, h, n, p = 100346.13, dni, dif, mintemp, maxtemp,
  lat, long, merid = 0, tz = 0, dst = 0)
}
\arguments{
\item{julian}{vector of julian days expressed as integers for every day for which \code{mintemp} and \code{maxtemp} are provided, as returned by function \code{\link[=julday]{julday()}}.}

\item{em}{an optional vector of hourly emissivities. If not provided, calculated from \code{h}, \code{n} and \code{p}.}

\item{h}{a vector of hourly specific humidities (\ifelse{html}{\out{kg kg<sup>-1</sup> }}{\eqn{kg kg^{-1}}}). Ignored if \code{em} provided.}

\item{n}{a vector of hourly fractional cloud cover values (range 0 - 1). Ignored if \code{em} provided.}

\item{p}{an optional vector of hourly atmospheric pressure values (Pa). Ignored if \code{em} provided.}

\item{dni}{a vector of hourly direct radiation values normal to the solar beam (\ifelse{html}{\out{MJ m<sup>-2</sup> hr<sup>-1</sup>}}{\eqn{MJ m^-2 hr^{-1}}}).}

\item{dif}{a vector of hourly diffuse radiation values (\ifelse{html}{\out{MJ m<sup>-2</sup> hr<sup>-1</sup>}}{\eqn{MJ m^-2 hr^{-1}}}).}

\item{mintemp}{a vector of daily minimum temperatures (ºC).}

\item{maxtemp}{a bector of daily maximum temperatures (ºC).}

\item{lat}{a single numeric value representing the latitude of the location for which hourly temperatures are required (decimal degrees, -ve south of equator).}

\item{long}{a single numeric value representing the longitude of the location for which hourly temperatures are required (decimal degrees, -ve west of Greenwich meridian).}

\item{merid}{an optional numeric value representing the longitude of the local time zone meridian (º) (0 for UK time).}

\item{tz}{an optional single  numeric value or vector of values specifying the time zones expressed as hours different from GMT for each day (-ve to west).}

\item{dst}{a single numeric value or vector of values representing the local summer time adjustment for each day (hours, e.g. +1 for BST).}
}
\value{
a vector of hourly temperatures (ºC).
}
\description{
\code{hourlytemp} is used to derive hourly temperatures from daily maxima and minima.
}
\details{
A warning is returned if any of following conditions are not
met. (1)  \code{h}, \code{n}, \code{dif} and \code{dct} differ in length.
(2) \code{julian}, \code{mintemp} and \code{maxtemp} differ in length. (3) E.g. \code{h} / 24 is not
an integer. (4) the length of e.g. \code{h} is not equal to the length of
e.g. \code{mintemp} x 24.
}
\examples{
jd <- julday(2010, 5 , c(1,2))
ht <- hourlytemp(jd, NA, microvars$humidity[1:48], microvars$cloudcover[1:48],
                 microvars$pressure[1:48], dni = microvars$dni[1:48],
                 dif = microvars$dif[1:48], c(7.5, 7.2), c(14.6, 15.2),
                 49.968, -5.216)
plot(ht ~ c(0:47),type="l", xlab = "Hour", ylab = "Temperature",
     main = paste("tmins:", 7.2, 7.5, "tmaxs:", 14.6, 15.2))
}
