% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othertools.R
\name{humidityconvert}
\alias{humidityconvert}
\title{Converts between different measures of humidity}
\usage{
humidityconvert(h, intype = "relative", tc = 20, p = 101300)
}
\arguments{
\item{h}{humidity value(s). Units as follows: specific humidity (\ifelse{html}{\out{kg kg<sup>-1</sup>}}{\eqn{kg kg^{-1}}}), absolute humidity (\ifelse{html}{\out{kg m<sup>-3</sup> }}{\eqn{kg m^{-3}}}), relative humidity (\\%), vapour pressure (kPa).}

\item{intype}{a character string description of the humidity type of \code{h}. One of "relative", "absolute" or "specific".}

\item{tc}{A numeric value specifying the temperature (ºC).}

\item{p}{An optional numeric value specifying the atmospheric pressure (Pa).}
}
\value{
a list of numeric humidity values with the following components:

\code{relative} relative humidity (\\%).

\code{absolute}  absolute humidity (\ifelse{html}{\out{kg m<sup>-3</sup> }}{\eqn{kg m^{-3}}}).

\code{specific} specific humidity (\ifelse{html}{\out{kg kg<sup>-1</sup> }}{\eqn{kg kg^{-1}}}).

\code{vapour_pressure} vapour pressure (kPa).
}
\description{
\code{humidityconvert} is used to convert between different measures of humidity, namely relative, absolute or specific. Vapour pressure is also returned.
}
\details{
This function converts between vapour pressure and specific,
relative and absolute humidity, based on sea-level pressure and
temperature. It returns a list of relative, absolute and specific
humidity and vapour pressure. If \code{intype} is unspecified, then \code{h} is
assumed to be relative humidity. If \code{p} is unspecified, pressure assumed
to be 101300, a typical value for sea-level pressure. If one or more of the
relative humidity values exceeds 100\\% a warning is given.
}
\examples{
humidityconvert(90, 'relative', 20)
humidityconvert(0.01555486, 'absolute', 20)
humidityconvert(0.01292172, 'specific', 20)
}
