% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{huss}
\alias{huss}
\title{A 1 km resolution dataset of interpolated daily specific humidity values}
\format{An array with 19 rows, 22 columns and 365 daily values.}
\source{
\url{https://eip.ceh.ac.uk/chess/}
}
\usage{
huss
}
\description{
A dataset containing daily specific humidity values (\ifelse{html}{\out{kg kg<sup>-1</sup> }}{\eqn{kg kg^{-1}}})
in 2010 for the area bounded by 160000, 182000, 11000, 30000  (xmin, xmax, ymin, ymax)
using the Ordance Survey GB Grid Reference system (CRS: +init=epsg:27700).
}
\keyword{datasets}
