% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solartools.R
\name{if_raster}
\alias{if_raster}
\title{Flexible conversion to SpatRaster object}
\usage{
if_raster(x, r)
}
\arguments{
\item{x}{an R object}

\item{r}{an R object}
}
\value{
if \code{r} is a SpatRaster, \code{x} is converted to a SpatRaster with the same attributes as \code{r}, otherwise returns \code{x}
}
\description{
\code{if_raster} is used to permit flexibility in the use of SpatRasts, matrices or arrays in many functions.
}
\examples{
r <- is_raster(dtm100m)
r1 <- if_raster(r, dtm100m)
r2 <- if_raster(r, r)
class(r1) # is a SpatRaster
class(r2) # is a matrix
}
