% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiationtools.R
\name{lai}
\alias{lai}
\title{Calculates Leaf Area Index}
\usage{
lai(red, nir, maxlai = 20)
}
\arguments{
\item{red}{a raster object, two-dimensional array or matrix of reflectance values in the red spectral band.}

\item{nir}{a raster object, two-dimensional array or matrix of reflectance values in the near-infrared spectral band.}

\item{maxlai}{an optional single numeric value representing the likely upper limit of Leaf Area Index to which all values are capped.}
}
\value{
a raster object or two-dimensional array of numeric values representing the total one-sided area of leaf tissue per unit ground surface area.
}
\description{
\code{lai} is used to calculate the total one-sided area of leaf tissue per unit ground surface area from the Normalized Difference Vegetation Index.
}
\details{
If \code{red} is a raster object, a raster object is returned. This function has been calibrated
using data derived from a small area of Cornwall only. It is strongly recommended that
locally calibrated values are obtained.
}
\examples{
leaf <- lai(aerial_image[,,3], aerial_image[,,4])
plot(if_raster(leaf, dtm1m), main = "Leaf area index")
}
\seealso{
function \code{\link[=lai_adjust]{lai_adjust()}} for calculated leaf area index values at specified heights above the ground.
}
