% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othertools.R
\name{laifromhabitat}
\alias{laifromhabitat}
\title{Derives leaf area index, leaf geometry and canopy height from habitat}
\usage{
laifromhabitat(habitat, lat, long, year, meantemp = NA, cvtemp = NA,
  rainfall = NA, cvrain = NA, wetmonth = NA)
}
\arguments{
\item{habitat}{a character string or numeric value indicating the habitat type. See \code{\link[=habitats]{habitats()}}}

\item{lat}{a single numeric value representing the mean latitude of the location for which the solar index is required (decimal degrees, -ve south of the equator).}

\item{long}{a single numeric value representing the mean longitude of the location for which the solar index is required (decimal degrees, -ve west of Greenwich meridian).}

\item{year}{the year for which data are required.}

\item{meantemp}{an optional numeric value of mean annual temperature (ºC) at reference height.}

\item{cvtemp}{an optional numeric value of the coefficient of variation in temperature (K per 0.25 days) at reference height.}

\item{rainfall}{an optional numeric value mean annual rainfall (mm per year)}

\item{cvrain}{an optional numeric value of the coefficient of variation in raifall (mm per 0.25 days) at reference height.}

\item{wetmonth}{an optional numeric value indicating which month is wettest (1-12).}
}
\value{
a list with the following items, (1) lai: hourly leaf area index values,
(2) x: the ratio of vertical to horizontal projections of leaf foliage,
(3) height: the heigbht of the canopy in metres and (4) obs_time: an object of
class POSIXlt of dates and times coressponding to each value of lai.
}
\description{
\code{laifromhabitat} generates an hourly dataset for an entire year of
leaf area index values, the ratio of vertical to horizontal projections of leaf
foliage and canopy height from habitat.
}
\details{
If no values of \code{meantemp}, \code{cvtemp}, \code{rainfall}, \code{cvrain} or \code{wetmonth} are
provided, values are obtained from \code{\link[=globalclimate]{globalclimate()}}. Variable \code{lai} is derived by fitting
a Gaussian curve parameters of which are climate and location-dependent. Functional
fits were calibrated using MODIS data (see https://modis.gsfc.nasa.gov/data/).
}
\examples{
lxh <- laifromhabitat("Deciduous broadleaf forest", 50, -5.2, 2015)
lxh$height
lxh$x
plot(lxh$lai ~ as.POSIXct(lxh$obs_time), type = "l", xlab = "Month",
     ylab = "LAI", ylim = c(0, 6))
}
\seealso{
\code{\link[=lai]{lai()}}, \code{\link[=leaf_geometry]{leaf_geometry()}}
}
