% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landsearatios}
\alias{landsearatios}
\title{Inverse-distance weighted land-sea ratios in 36 directions}
\format{
An array with 187 rows, 214 columns and 36 directional values.
}
\usage{
landsearatios
}
\description{
A 100m resolution dataset of nverse-distance weighted land-sea ratios, as
produced by function \code{\link[=invls]{invls()}} in each of 36 directions (0º, 10º..350º) for
the area bounded by 160000, 181400, 11300, 30000  (xmin, xmax, ymin, ymax)
using the Ordance Survey GB Grid Reference system (CRS: +init=epsg:27700).
}
\keyword{datasets}
