% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othertools.R
\name{lapserate}
\alias{lapserate}
\title{Calculates the moist adiabatic lapse rate}
\usage{
lapserate(tc, h, p = 101300)
}
\arguments{
\item{tc}{a single numeric value, SpatRaster object, two-dimensional array or matrix of temperature (ºC).}

\item{h}{a single numeric value, SpatRaster object, two-dimensional array or matrix of specific humidity (\ifelse{html}{\out{kg kg<sup>-1</sup> }}{\eqn{kg kg^{-1}}}).}

\item{p}{an optional single numeric value, SpatRaster object, two-dimensional array or matrix of atmospheric pressure (Pa).}
}
\value{
the lapse rate (\ifelse{html}{\out{º m<sup>-1</sup> }}{\eqn{ º m^{-1}}}).
}
\description{
\code{lapserate} is used to calculate changes in temperature with height.
}
\details{
if tc is a SpatRaster, a SpatRaster object is returned. This function calculates the
theoretical lapse rate. Environmental lapse rates can vary due to winds.
}
\examples{
lapserate(20, 0) * 1000 # dry lapse rate per km
h <- humidityconvert(100, intype = "relative", 20)
lapserate(20, h$specific) * 1000 # lapse rate per km when relative humidity is 100\%
}
