% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiationtools.R
\name{leaf_geometry}
\alias{leaf_geometry}
\title{Calculates leaf orientation}
\usage{
leaf_geometry(veghgt, maxx = 8)
}
\arguments{
\item{veghgt}{a raster object, two-dimensional array or matrix of vegetation heights (m).}

\item{maxx}{a theoretical upper limit for the ratio of vertical to horizontal projections of leaf foliage, to which all values are capped.}
}
\value{
a raster object or a two-dimensional array of numeric values representing the
ratio of vertical to horizontal projections of leaf foliage. The output tends towards
zero as vegetation is more vertically orientated and maxx as it is more horizontally
orientated.
}
\description{
\code{leaf_geometry} is used to calculates the ratio of vertical to horizontal projections of leaf foliage.
}
\details{
Under vegetated canopies, canopy transmission not only decreases with canopy cover but is
also affected by leaf structure. At low solar angles, radiation is lower when leaves are
more vertically oriented and \code{leaf_geometry} is hence used to calculate an approximate  factor
indicating the degree to which vegetation is vertically orientated based on the premise that
shorter vegetation is more likely to have more vertically orientated leaves.
If \code{veghgt} is a raster object, a raster object is returned.
This function has been calibrated using data derived from a small area of
Cornwall only. It is strongly recommended that locally calibrated values
units of x, y and z are identical. Use \code{\link[=projectRaster]{projectRaster()}} to convert the
projection to a Universal Transverse Mercator type projection system.
}
\examples{
library(raster)
x <- leaf_geometry(veg_hgt)
plot(x, main = "Leaf geometry")
}
