% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiationtools.R
\name{longwaveveg}
\alias{longwaveveg}
\title{Calculates net longwave radiation below canopy}
\usage{
longwaveveg(h, tc, p = 101300, n, x, fr, svv = 1, albc = 0.23)
}
\arguments{
\item{h}{a single numeric value, raster object, two-dimensional array or matrix of specific humidities (\ifelse{html}{\out{kg kg<sup>{-1}</sup> }}{\eqn{kg kg^{-1}}}).}

\item{tc}{a single numeric value, raster object, two-dimensional array or matrix of temperatures (ºC).}

\item{p}{an optional single numeric value, raster object, two-dimensional array or matrix of sea-level pressures (Pa).}

\item{n}{a single numeric value, raster object, two-dimensional array or matrix of fractional cloud cover (range 0 - 1).}

\item{x}{a raster object, two-dimensional array or matrix of numeric values representing the ratio of vertical to horizontal projections of leaf foliage as returned by \code{\link[=leaf_geometry]{leaf_geometry()}}.}

\item{fr}{a raster object, two-dimensional array or matrix of fractional canopy cover as returned by \code{\link[=canopy]{canopy()}}.}

\item{svv}{an optional raster object, two-dimensional array or matrix of values representing the proportion of isotropic radiation received by a surface partially obscured by topography relative to the full hemisphere underneath vegetation as returned by \code{\link[=skyviewveg]{skyviewveg()}}.}

\item{albc}{an optional single value, raster object, two-dimensional array or matrix of values representing the albedo(s) of the vegetated canopy as returned by \code{\link[=albedo2]{albedo2()}}.}
}
\value{
a single numeric value, raster object or two-dimensional array of values representing net longwave radiation (MJ per metre squared per hour).
}
\description{
\code{longwaveveg} is used to calculate a high-resolution dataset of the net longwave radiation flux density emmited from the Earth, accounting for canopy effects.
}
\details{
If \code{svv} is a raster object, a raster object is returned.
If no values for \code{p} are provided, a default value of 101300 Pa, typical of
sea-level pressure, is assumed. If no value for \code{albc} is provided, then
the entire area is assumed to have a default value of 0.23, typical of
well-watered grass. If single values of \code{h}, \code{tc}, \code{p}, \code{n} or
\code{albc} are given, then the entire area is assumed to have the same values.
If no value for \code{svv} is provided then the entire hemisphere is assumed to
be in view.
}
\examples{
# =================================
# Extract data for 2010-05-24 11:00
# =================================
h <- microvars$humidity[564]
p <- microvars$pressure[564]
n <- microvars$cloudcover[264]
tcr <- raster(temp100[,,564], xmn = 169000, xmx = 170000, ymn = 12000, ymx = 13000)
tc <- resample(tcr, dtm1m) # Resample temperature raster to 1m
# =========================
# calculate input variables
# =========================
x <- leaf_geometry(veg_hgt)
l <- lai(aerial_image[,,3], aerial_image[,,4])
l <- lai_adjust(l, veg_hgt)
svv <- skyviewveg(dtm1m, l, x)
fr <- canopy(l, x)
alb <- albedo(aerial_image[,,1], aerial_image[,,2], aerial_image[,,3],
             aerial_image[,,4])
albc <- albedo2(alb, fr, ground = FALSE)
# =====================================
# calculate and plot longwave radiation
# =====================================
netlong1m <-longwaveveg(h, tc, p, n, x, fr, svv, albc)
nlr <- mask(netlong1m, dtm1m)
plot(nlr, main = "Net longwave radiation")
}
\seealso{
The function \code{\link[=longwavetopo]{longwavetopo()}} returns the net longwave radiation above vegetation.
The function \code{\link[=humidityconvert]{humidityconvert()}} can be used to derive specific humidy from other meaures of humidity.
}
