% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools.R
\name{microclimaforNMR}
\alias{microclimaforNMR}
\title{Extract NCEP data and run microclima for use with NicheMapR}
\usage{
microclimaforNMR(lat, long, dstart, dfinish, l, x, hourlydata = NA,
  dailyprecip = NA, dem = NA, albr = 0.15, resolution = 30, zmin = 0,
  slope = NA, aspect = NA, windthresh = 4.5, emthresh = 0.78,
  reanalysis2 = TRUE)
}
\arguments{
\item{lat}{the latitude (in decimal degrees) of the location for point data are required.}

\item{long}{the longitude (in decimal degrees) of the location for point data are required.}

\item{dstart}{start date as character of time period required in format DD/MM/YYYY}

\item{dfinish}{end date as character of time period required in format DD/MM/YYYY}

\item{l}{a single numeric value of the leaf area index for location (see details).}

\item{x}{a single numeric value representing the ratio of vertical to horizontal
projections of leaf foliage for the location (see details).}

\item{hourlydata}{an optional data frame of ourly weather and radiation data as returned
by function \code{\link[=hourlyNCEP]{hourlyNCEP()}}. Function called if data not provided.}

\item{dailyprecip}{an optional data frame of ourly weather and radiation data as returned
by function \code{\link[=dailyprecipNCEP]{dailyprecipNCEP()}}. Function called if data not provided.}

\item{dem}{optional raster object of elevations covering the location for which point
data are required. If not provided, one is downloaded from the registry of Open
Data on AWS.}

\item{albr}{albedo of ground surface from which radiation is reflected (see details)}

\item{resolution}{the resolution (in metres) of the dem used for downscaling (see details).}

\item{zmin}{assumed sea-level height. Values below this are set to zmin (see details).}

\item{slope}{an optional slope value for the location in decimal degrees. If not specified, then
the slope is calculated from the retrieved dem.}

\item{aspect}{an optional aspect value for the location in decimal degrees. If not specified, then
the slope is calculated from the retrieved dem.}

\item{windthresh}{an optional threshold wind value (m /s) above which cold air drainage is
assumed not to occur}

\item{emthresh}{an optional threshold emissivity value above which cold air drainage is
assumed not to occur}

\item{reanalysis2}{Logical. Should data be obtained from the Reanalysis II dataset (default) or
from Reanalysis I (data prior to 1979).}
}
\value{
a list with the following objects:

(1) hourlydata: a data frame of hourly climate data from NCEP,
as returned by \code{\link[=hourlyNCEP]{hourlyNCEP()}}.

(2) hourlyradwind: A data frame with seven columns:
(i) \code{swrad}: total downward shortwave radiation as a funcion of slope, aspect and
canopy shading (\ifelse{html}{\out{MJ m<sup>-2</sup> hr<sup>-1</sup>}}{\eqn{MJ m^-2 hr^{-1}}})
as returned by \code{\link[=shortwaveveg]{shortwaveveg()}}.(ii) \code{skyviewfact}: a topographic sky view correction factor
as returned by \code{\link[=skyviewveg]{skyviewveg()}}. (iii) \code{canopyfact} a canopy shading factor as returned
by \code{\link[=canopy]{canopy()}}. (iv) \code{whselt}: a ground level wind shelter coefficient as returned by
\code{\link[=windcoef]{windcoef()}}. (v) \code{windspeed}: wind speed (m /s) at 1 m, corrected for topographic
sheltering. (vi) \code{slope}: the slope (decimal degrees) of the location specified
by \code{lat} and \code{long}. Either the same as that input, or determined from \code{dem}.
(vii) \code{aspect}: the aspect (decimal degrees) of the location specified by \code{lat} and
\code{long}. Either the same as that input, or determined from \code{dem}.

(3) \code{tref}: a dataframe with five columns (i) \code{tref} the reference temperature, derived from
NCEP data. (ii) \code{elev} the elevation of the location. (iii) \code{elevncep}: the mean elevation
of the 2.5º grid cell corresponding to that from which NCEP data were derived. (iv)
\code{telev} the expected difference in temperature due to elevation, calculated by
applying \code{\link[=lapserate]{lapserate()}}. (v) \code{tcad} the expected difference in temperature due to cold
air drainage effects, determined using \code{\link[=pcad]{pcad()}} and \code{\link[=cadconditions]{cadconditions()}}.

(4) \code{dailyprecip} a vector of daily rainfalls derived from NCEP.
}
\description{
Extract NCEP data and run microclima for use with NicheMapR
}
\details{
The package \code{NicheMapR} (Kearney & Porter 2016), includes a suite of
programs for mechanistic modelling of heat and mass exchange and provides a
time-series of hourly estimates of above- and below‐ground conditions. It has the
significant advantage over \code{microclima} in that it works entirely from first
principles does not require measurements of temperature for calibration, but does
not explicitly model elevation and cold-air drainage effects. The slope, aspect
and canopy effects must be specified by the user, and their effect on radiation
are more simplistically modelled than by \code{microclima}.
}
\examples{
mnr <- microclimaforNMR(50, -5.2, '15/01/2015', '15/02/2015', 1, 1)
head(mnr$hourlydata)
head(mnr$hourlyradwind)
head(mnr$tref)
head(mnr$dailyprecip)
}
\references{
Kearney MR,  Porter WP (2016) NicheMapR – an R package for biophysical
modelling: the microclimate model. Ecography 40: 664-674.
}
\seealso{
Function \code{\link[=get_dem]{get_dem()}} for retrieving digital elevation data and
\code{\link[=hourlyNCEP]{hourlyNCEP()}} for retrieving and processing hourly data from NCEP.
}
