% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{microvars}
\alias{microvars}
\title{Climate variables for  May 2010.}
\format{
A data frame with 744 rows and 9 variables:
\describe{
\item{dni}{Direct Normal Radiation (\ifelse{html}{\out{MJ m<sup>-2</sup> hr<sup>-1</sup>}}{\eqn{MJ m^{-2} hr^{-1}}})}
\item{dif}{Diffuse Radiation (\ifelse{html}{\out{MJ m<sup>-2</sup> hr<sup>-1</sup>}}{\eqn{MJ m^{-2} hr^{-1}}})}
\item{humidity}{Specific humidity (\ifelse{html}{\out{kg kg<sup>-1</sup> }}{\eqn{kg kg^{-1}}})}
\item{pressure}{Sea-level pressure (Pa)}
\item{cloudcover}{Fractional cloud cover}
\item{year}{Year (AD)}
\item{month}{Numeric month}
\item{day}{Day of month}
\item{hour}{Hour of day}
}
}
\usage{
microvars
}
\description{
A dataset containing hourly coarse-resolution climate variables in May 2010.
}
\keyword{datasets}
