% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{modis}
\alias{modis}
\title{A ~500 m resolution dataset of MODIS-derived albedo values}
\format{A matrix with 18 rows and 36 columns.}
\source{
\url{https://lpdaac.usgs.gov/}
}
\usage{
modis
}
\description{
A dataset ground surface albedo for the area bounded by
169000, 170000, 12000, 13000  (xmin, xmax, ymin, ymax) using the Ordance
Survey GB Grid Reference system (CRS: +init=epsg:27700) derived from
Moderate Resolution Imaging Spectroradiometer (MODIS) imagery
}
\keyword{datasets}
