% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{netlong1m}
\alias{netlong1m}
\title{A one metre resolution matrix of net longwave radiation}
\format{
A matrix with 1000 rows and 1000 columns.
}
\usage{
netlong1m
}
\description{
A dataset containing net longwave radiation values (\ifelse{html}{\out{MJ m<sup>-2</sup> hr<sup>-1</sup>}}{\eqn{MJ m^-2 hr^{-1}}})
for 2010-05-24 11:00 across the area bounded by 169000, 170000, 12000, 13000
(xmin, xmax, ymin, ymax) using the Ordance Survey GB Grid Reference
system (CRS: +init=epsg:27700), as produced by \code{\link[=longwaveveg]{longwaveveg()}}
}
\keyword{datasets}
