% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{netshort100m}
\alias{netshort100m}
\title{A 100 metre resolution matrix of net shortwave radiation}
\format{
A matrix with 187 rows and 214 columns.
}
\usage{
netshort100m
}
\description{
A dataset containing net shortwave radiation values (\ifelse{html}{\out{MJ m<sup>-2</sup> hr<sup>-1</sup>}}{\eqn{MJ m^{-2} hr^{-1}}})
for 2010-05-24 11:00 GMT across the area bounded by 160000, 181400, 11300, 30000
(xmin, xmax, ymin, ymax) using the Ordance Survey GB Grid Reference
system (CRS: +init=epsg:27700), as produced by \code{\link[=shortwavetopo]{shortwavetopo()}}
}
\keyword{datasets}
