% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadtools.R
\name{pcad}
\alias{pcad}
\title{Calculates cold air drainage potential}
\usage{
pcad(dem, basins, fa, tc, h, p = 101300, out = "cadp")
}
\arguments{
\item{dem}{a raster object of elevation (m).}

\item{basins}{a raster object, two-dimensional array or matrix with basins numbered as integers as returned by \code{\link[=basindelin]{basindelin()}}.}

\item{fa}{a raster object of accumulated flow, as returned by \code{\link[=flowacc]{flowacc()}}}

\item{tc}{a single numeric value, raster object, two-dimensional array or matrix of values with the dimensions as \code{dem} of temperature (ºC).}

\item{h}{a single numeric value, raster object, two-dimensional array or matrix of values with the dimensions as \code{dem} of specific humidity (\ifelse{html}{\out{kg kg<sup>-1</sup> }}{\eqn{kg kg^{-1}}}).}

\item{p}{an optional single numeric value, raster object, two-dimensional array or matrix of values with the dimensions as \code{dem} of atmospheric pressure (Pa).}

\item{out}{specifies the type of output to provide (see details). Possible values are "cadp", "tempdif" and "pflow".}
}
\value{
a raster object:
\describe{
\item{if \code{out = "tempdif"}}{the expected temperature difference (ºC).}
\item{if \code{out = "pflow"}}{the accumulated flow logairthmically transformed expressed as a proportion of the maximum in each basin.}
\item{if \code{out = "cadp"} (the default)}{the expected temperature difference x the proportion of accumulated flow.}
}
}
\description{
\code{ pcad} calculates the expected temperature differences resulting from cold air
drainage.
}
\details{
To derive expected temperature differences, \code{pcad} calculates the
difference difference in elevation between each point and the highest in the
basin and multiplies this by the lapse rate. Accumulated flow is calculated
using \code{\link[=flowacc]{flowacc()}}, logarithmically transformed and expressed as proportion of
the maximum in each basin as an indication of each locations potential to
experience cold air drainage. Cold air flow is possible over shallow
boundaries, \code{basins} is best derived using \code{\link[=basinmerge]{basinmerge()}}.
Warning: function is quite slow on large datasets.
}
\examples{
library(raster)
basins <- basinmerge(dtm100m, basins100m, 2)
h <- humidityconvert(50, intype = "relative", 20)$specific
fa <- flowacc(dtm100m, basins)
cp1 <- pcad(dtm100m, basins, fa, 20, h)
cp2 <- pcad(dtm100m, basins, fa, 20, h, out = "tempdif")
cp3 <- pcad(dtm100m, basins, fa, 20, h, out = "pflow")
par(mfrow=c(1, 3))
plot(cp3, main = "Accumulated flow proportion")
plot(cp2, main = "Expected temperature difference")
plot(cp1, main = "Cold air drainage potential")
}
