% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools.R
\name{runauto.ncep}
\alias{runauto.ncep}
\title{Function for automatically generating microclimate surfaces for anywhere in the word}
\usage{
runauto.ncep(r, dstart, dfinish, hgt = 0.05, l, x, habitat = NA,
  use.raster = FALSE, coastal = TRUE, r.is.dem = TRUE, save = 0,
  albg = 0.15, albr = 0.15, albc = 0.23, mesoresolution = 100,
  zmin = 0, slope = NA, aspect = NA, windthresh = 4.5,
  emthresh = 0.78, reanalysis2 = TRUE, steps = 8,
  plot.progress = TRUE, continuous = TRUE, summarydata = TRUE,
  save.memory = FALSE)
}
\arguments{
\item{r}{a raster object defining the extent and resolution for which microclimate
temperature data are required.}

\item{dstart}{start date as character of time period required in format DD/MM/YYYY}

\item{dfinish}{end date as character of time period required in format DD/MM/YYYY}

\item{hgt}{the height (in m) above or below ground for which temperature estimates
are required. If negative, below-ground temperatures are derived (range -2 to 2).}

\item{l}{an optional single numeric value, vector, raster, matrix or array of leaf area
index values. If a single numeric value, the same leaf area is assumed at all locations
and times. If a vector, the same leaf area is assumed at all locations, but leaf area is
assumed ot vary through time. If a matrix or raster, leaf area is assumed to vary by
location, but is assumed static in time. If an array, variable leaf areas in time and
space are assumed.}

\item{x}{a single numeric value, matrix or raster representing the ratio of vertical to horizontal projections of foliage as,
for example, returned by \code{\link[=leaf_geometry]{leaf_geometry()}}. if a single numeric value, the same value of \code{x}
is assumed at all locations. Varying \code{x} through time is not supported.}

\item{habitat}{a character string, numeric value or matrix or raster of numeric
values of habitat type(s). See \code{\link[=habitats]{habitats()}}. Ignored if l or x are provided. If a single
numeric value, the same habitat is assumed at all locations. Varying habitats in time are not
supported.}

\item{use.raster}{optional logical value indicating whether to use \code{r} in the
calculation of coastal effects. If used, NAs or values corresponding to \code{zmin}
must represent sea.}

\item{coastal}{optional logical value indicating whether or not to calculate
coastal effects.}

\item{r.is.dem}{optional logical value indicating whether 'r' is a digital
elevation dataset used for calculating microclimatic effects. If FALSE, then a
dem is downloaded.}

\item{save}{optional integer: don't save forcing data (0), save the forcing data (1)
or read previously saved data (2).}

\item{albg}{an optional single value, raster object, two-dimensional array or
matrix of values representing the albedo(s) of the ground as returned by \code{\link[=albedo2]{albedo2()}}.}

\item{albr}{albr an optional single value, raster object, two-dimensional array
or matrix of values representing the albedo(s) of adjacent surfaces as returned
by \code{\link[=albedo_reflected]{albedo_reflected()}}.}

\item{albc}{an optional single value, raster object, two-dimensional array or
matrix of values representing the albedo(s) of the vegetated canopy as returned
by \code{\link[=albedo2]{albedo2()}}.}

\item{mesoresolution}{optional numeric value indicating resolution of the dem used for
modelling mesoclimate.}

\item{zmin}{assumed sea-level height. Values below this are set to zmin (see details).}

\item{slope}{an optional slope value for the location in decimal degrees. If not specified, then
the slope is calculated from the retrieved dem.}

\item{aspect}{an optional aspect value for the location in decimal degrees. If not specified, then
the slope is calculated from the retrieved dem.}

\item{windthresh}{an optional threshold wind value (m /s) above which cold air drainage is
assumed not to occur.}

\item{emthresh}{an optional threshold emissivity value above which cold air drainage is
assumed not to occur.}

\item{reanalysis2}{Optional logical. Should data be obtained from the Reanalysis II dataset (default) or
from Reanalysis I (data prior to 1979).}

\item{steps}{an optional integer. Coastal effects are calculated in specified
directions upwind. Steps defines the total number of directions used. If the
default 8 is specified, coastal effects are calculated at 45º intervals.}

\item{plot.progress}{an optional logical indicating whether to produce plots to track progress.}

\item{continuous}{an optional logical value indicating whether to treat wind speed as a continuous variable}

\item{summarydata}{an optional logical indicating whether to calculate summary data
(frost hours and maximum, minimum and mean temperature) for each pixel and return these to the output.}

\item{save.memory}{An optional logical indicatign whether to save
memory by storing temperature x 1000 as an integer values.}

\item{dailyprecip}{Optional data.frame of daily precipitation data as returned by \code{\link[=dailyprecipNCEP]{dailyprecipNCEP()}}}
}
\value{
a list with the following objects:
(1) temps: an array of temperatures for each pixel of r and hour of the time sequence.
(2) e: an extent object given the extent of \code{temps}. Generally the same as \code{raster::extent(r)}
though note that edge cells are NA as slopes cannot be calculate for these cells.
(3) units: the units of \code{temps}. Either deg C or dec C * 1000 if \code{save.memory} is TRUE.
(4) tmax: If \code{summarydata} is TRUE, a matrix of maximum temperatures
(5) tmin: If \code{summarydata} is TRUE, a matrix of minimum temperatures
(6) tmean: If \code{summarydata} is TRUE, a matrix of mean temperatures
(7) frosthours: If \code{summarydata} is TRUE, a matrix of hours below 0 deg C.
}
\description{
This function generating microclimate temperature surfaces for anywhere
in the word. It first downloads coarse-resolution climate and radiationdata from The
NCEP-NCAR or NCEP–DOE Atmospheric Model Intercomparison Project (Kanamitso et al
2002) and interpolates these data to hourly. It calculates mesoclimatic effects and derives parameters for fitting the
microclimate model using the \code{NicheMapR} package (Kearny & Porter 2016). Using digital
elevation data either downloaded or provided by the user, and canopy
characteristics either specified by the user, or derived from habitat,
it runs the microclimate model in hourly timesteps to generate an array of temperatures.
}
\examples{
library(raster)
require(NicheMapR)
# Get dem for Pico, Azores
r <- get_dem(lat = 38.467429, long = -28.398995, resolution = 30)
plot(r)
# Takes ~ c. 15 minutes to run
temps <- runauto.ncep(r, "10/06/2010", "15/06/2010", hgt = 0.1, l = NA, x = NA,
                     habitat = "Barren or sparsely vegetated")
mypal <- colorRampPalette(c("darkblue", "blue", "green", "yellow", "orange",
                          "red"))(255)
meantemp <- temps$tmean
extent(meantemp) <- temps$e
par(mfrow = c(1, 1))
plot(meantemp, main = "Mean temperature", col = mypal)
# Interactive 3D plot
require(plotly)
zrange<-list(range = c(0, 3000))
plot_ly(z = ~is_raster(r)) \%>\%
add_surface(surfacecolor = ~is_raster(meantemp)) \%>\%
layout(scene = list(zaxis = zrange))
}
\references{
Kearney MR,  Porter WP (2016) NicheMapR – an R package for biophysical
modelling: the microclimate model. Ecography 40: 664-674.

Kanamitsu M, Ebisuzaki W, Woollen J, Yang SK, Hnilo JJ, Fiorino M, Potter GL (2002)
Ncep–doe amip-ii reanalysis (r-2). Bulletin of the American Meteorological Society 83: 1631-1644.
}
