% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiationtools.R
\name{shortwavetopo}
\alias{shortwavetopo}
\title{Downscales shortwave radiation accounting for topographic effects}
\usage{
shortwavetopo(dni, dif, julian, localtime, lat = NA, long = NA,
  dtm = array(0, dim = c(1, 1)), slope = NA, aspect = NA, svf = 1,
  alb = 0.23, albr = 0.23, ha = 0, res = 100, merid = 0, dst = 0,
  shadow = TRUE, component = "sw")
}
\arguments{
\item{dni}{a single numeric value, raster object, two-dimensional array or matrix of coarse-resolution direct radiation perpendicular to the solar beam (\ifelse{html}{\out{MJ m<sup>-2</sup> hr<sup>-1</sup>}}{\eqn{MJ m^{-2} hr^{-1}}}).}

\item{dif}{a single numeric value, raster object, two-dimensional array or matrix of diffuse radiation horizontal ot the surface (\ifelse{html}{\out{MJ m<sup>-2</sup> hr<sup>-1</sup>}}{\eqn{MJ m^{-2} hr^{-1}}}).}

\item{julian}{a single integer representing the Julian as returned by \code{\link[=julday]{julday()}}.}

\item{localtime}{a single numeric value representing local time (decimal hour, 24 hour clock).}

\item{lat}{a single numeric value representing the mean latitude of the location for which downscaled radiation is required (decimal degrees, -ve south of equator).}

\item{long}{a single numeric value representing the mean longitude of the location for which downscaled radiation is required (decimal degrees, -ve west of Greenwich meridian).}

\item{dtm}{an optional raster object, two-dimensional array or matrix of elevations (m), orientated as if derived using \code{\link[=is_raster]{is_raster()}}. I.e. \code{[1, 1]} is the NW corner.}

\item{slope}{a single value, raster object, two-dimensional array or matrix of slopes (º). If an array or matrix, then orientated as if derived using \code{\link[=is_raster]{is_raster()}}. I.e. \code{[1, 1]} is the NW corner.}

\item{aspect}{a single value, raster object, two-dimensional array or matrix of aspects (º). If an array or matrix, then orientated as if derived using \code{\link[=is_raster]{is_raster()}}. I.e. \code{[1, 1]} is the NW corner.}

\item{svf}{an optional single value, raster object, two-dimensional array or matrix of values representing the proportion of isotropic radiation received by a partially obscured surface relative to the full hemisphere as returned by \code{\link[=skyviewtopo]{skyviewtopo()}}.}

\item{alb}{an optional single value, raster object, two-dimensional array or matrix of surface albedo(s) (range 0 - 1) derived using \code{\link[=albedo]{albedo()}} or \code{\link[=albedo_adjust]{albedo_adjust()}}.}

\item{albr}{an optional single value, raster object, two-dimensional array or matrix of values of albedo(s) of adjacent surfaces (range 0 - 1) as returned by \code{\link[=albedo_reflected]{albedo_reflected()}}.}

\item{ha}{an optional raster object, two-dimensional array or matrix of values representing the mean slope to the horizon (decimal degrees) of surrounding surfaces from which radiation is reflected for each cell of \code{dtm} as returned by \code{\link[=mean_slope]{mean_slope()}}.}

\item{res}{a single numeric value representing the spatial resolution of \code{dtm} (m).}

\item{merid}{an optional single numeric value representing the longitude (decimal degrees) of the local time zone meridian (0 for UK time).}

\item{dst}{an optional numeric value representing the local summer time adjustment (hours in 24 hour clock)  (+1 for BST, +0 for GMT).}

\item{shadow}{an optional logical value indicating whether topographic shading should be considered (False = No, True = Yes).}

\item{component}{an optional character string of the component of radiation to be returned. One of "sw" (net shortwave radiation, i.e. accounting for albedo), "sw2" (total incoming shortwave radiation), "dir" (direct), "dif" (diffuse), "iso" (isotropic diffuse), "ani" (anistopic diffuse), "ref" (reflected).}
}
\value{
If component is "sw", a raster object or two-dimensional array of numeric values representing net shortwave radiation (MJ m^-2 hr^-1).

If component is "sw2", a raster object or two-dimensional array of numeric values representing total incoming shortwave radiation (MJ m^-2 hr^-1).

If component is "dir", a raster object or two-dimensional array of numeric values representing direct shortwave radiation (MJ m^-2 hr^-1).

If component is "dif", a raster object or two-dimensional array of numeric values representing diffuse shortwave radiation (MJ m^-2 hr^-1).

If component is "iso", a raster object or two-dimensional array of numeric values representing isotropic diffuse shortwave radiation (MJ m^-2 hr^-1).

If component is unspecified, then the default "sw" is returned.
}
\description{
\code{shortwavetopo} is used to downscale components of the flux density of
shortwave radiation received at the surface of the Earth using a
high-resolution digital elevation dataset, ignoring canopy effects.
}
\details{
If \code{slope} is unspecified, and \code{dtm} is a raster, \code{slope} and \code{aspect} are calculated from
the raster. If \code{slope} is unspecified, and \code{dtm} is not a raster, the slope and aspect
are set to zero. If \code{lat} is unspecified, and \code{dtm} is a raster with a coordinate reference
system defined, \code{lat} and \code{long} are calculated from the raster. If \code{lat} is unspecified,
and \code{dtm} is not a raster, or a raster without a coordinate reference system defined, an
error is returned. If \code{dtm} is specified, then the projection system used must be such that
units of x, y and z are identical. Use \code{\link[=projectRaster]{projectRaster()}} to convert the projection to a
Universal Transverse Mercator type projection system. If \code{dtm} is a raster object, a raster
object is returned. If \code{dtm} is a raster object, a raster object is returned. If \code{dni} or \code{dif} are raster
objects, two-dimensional arrays or matrices, then it is assumed that they have been
derived from coarse-resolution data by interpolation, and have the same extent as \code{dtm}.
If no value for \code{ha} is provided, the mean slope to the horizon is assumed
to be 0. If no value for \code{svv} is provided, then the entire hemisphere is
assumed to be in view. If no value for \code{svf} is provided, then the entire
hemisphere is assumed to be in view. If values of \code{alb} and \code{albr} are not
specified, then a default value of 0.23, typical of well-watered grass is
assumed. If single values of \code{alb} and \code{albr} are given, then the entire
area is assumed to have the same albedo. If \code{dtm} is specified, then the
projection system used must be such that the units of x, y and z are
identical. Use \code{\link[=projectRaster]{projectRaster()}} to convert the projection to a Universal
Transverse Mercator type projection system. If no value for \code{dtm} is
provided, radiation is downscaled by deriving values on the inclined
surfaces specified in \code{slope} and \code{aspect} and topographic shadowing is
ignored. If single values are provided for \code{slope} and \code{aspect} single
values of components of shortwave radiation for an inclined surface are
returned. Only single values of \code{lat} and \code{long} are taken as inputs. If
\code{dtm} covers a large extent, the \code{dtm} is best divided into blocks and
seperate calculations performed on each block. Since horizon angles,
topographic shading and sky view correction factors may be influenced by
locations beyond the extent of \code{dtm}, it is best to ensure \code{dtm} covers a
larger extent than that for which radiation values are needed, and to
ensure sub-divided blocks overlap in extent. Calculations are faster if values
for all inputs are provided.
}
\examples{
# =================================
# Extract data for 2010-05-24 11:00
# =================================
dni <-dnirad[,,3444]
dif <-difrad[,,3444]
# ===========================
# Resample to 100m resolution
# ===========================
dnir <- raster(dni, xmn = -5.40, xmx = -5.00, ymn = 49.90, ymx = 50.15)
difr <- raster(dif, xmn = -5.40, xmx = -5.00, ymn = 49.90, ymx = 50.15)
crs(dnir) <- '+init=epsg:4326'
crs(difr) <- '+init=epsg:4326'
dnir <- projectRaster(dnir, crs = "+init=epsg:27700")
difr <- projectRaster(difr, crs = "+init=epsg:27700")
dni <- resample(dnir, dtm100m)
dif <- resample(difr, dtm100m)
sv <- skyviewtopo(dtm100m)
jd <- julday(2010, 5, 24)
ha <- mean_slope(dtm100m)
# ================================================================
# Calculate and plot net shortwave radiation for 2010-05-24 11:00
# ================================================================
netshort100m <- shortwavetopo(dni, dif, jd, 11, dtm = dtm100m,
                              svf = sv, ha = ha)
plot(mask(netshort100m, dtm100m),
     main = "Net shortwave radiation")
}
\seealso{
Function \code{\link[=shortwaveveg]{shortwaveveg()}} returns net shortwave radiation below a canopy.
}
