% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools.R
\name{siflat}
\alias{siflat}
\title{Calculates the solar index for a flat surface}
\usage{
siflat(localtime, lat, long, julian, merid = round(long/15, 0) * 15, dst = 0)
}
\arguments{
\item{localtime}{local time (decimal hour, 24 hour clock).}

\item{lat}{latitude of the location for which the solar altitude is required (decimal degrees, -ve south of the equator).}

\item{long}{longitude of the location for which the solar altitude is required (decimal degrees, -ve west of Greenwich meridian).}

\item{julian}{Julian day expressed as an integer as returned by \code{\link[=julday]{julday()}}.}

\item{merid}{an optional numeric value representing the longitude (decimal degrees) of the local time zone meridian (0 for GMT). Default is \code{round(long / 15, 0) * 15}}

\item{dst}{an optional numeric value representing the time difference from the timezone meridian (hours, e.g. +1 for BST if \code{merid} = 0).}
}
\value{
a numeric value representing the solar altitude (º).
}
\description{
Calculates the solar index for a flat surface
}
\examples{
# solar index at noon GMT on 21 June 2010, Porthleven, Cornwall
jd <- julday (2010, 6, 21) # Julian day
siflat(12, 50.08, -5.31, jd)

}
