% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solartools.R
\name{solartime}
\alias{solartime}
\title{Calculates the solar time}
\usage{
solartime(localtime, long, julian, merid = 0, dst = 0)
}
\arguments{
\item{localtime}{local time (decimal hour, 24 hour clock).}

\item{long}{longitude of the location for which the solar time is required (decimal degrees, -ve west of Greenwich meridian).}

\item{julian}{Julian day expressed as an integer as returned by \code{\link[=julday]{julday()}}.}

\item{merid}{an optional numeric value representing the longitude (decimal degrees) of the local time zone meridian (0 for UK time).}

\item{dst}{an optional numeric value representing the local summer time adjustment (hours, e.g. +1 for BST).}
}
\value{
the solar time. I.e. the times that would be measured by a sundial (hours).
}
\description{
\code{solartime} is used to calculate the solar time. I.e. the time that would be measured by a sundial.
}
\details{
‘solartime’ accounts for two factors: firstly, east or west component of the analemma, namely
the angular offset of the Sun from its mean position on the celestial sphere as viewed from
Earth due the eccentricity of the Earth's orbit and the obliquity due to tilt of the Earth's
rotational axis. These two factors have different wavelengths, amplitudes and phases,
that vary over geological timescales. The equations used here are those derived by Milne.
}
\examples{
jd <- julday (2010, 6, 21) # Julian day
solartime(12, -5, jd) # solartime at noon on 21 June 2010, 5ºW
}
