% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solartools.R
\name{solazi}
\alias{solazi}
\title{Calculates the solar azimuth}
\usage{
solazi(localtime, lat, long, julian, merid = 0, dst = 0)
}
\arguments{
\item{localtime}{local time (decimal hour, 24 hour clock).}

\item{lat}{latitude of the location for which the solar azimuth is required (decimal degrees, -ve south of the equator).}

\item{long}{longitude of the location for which the solar azimuth is required (decimal degrees, -ve west of Greenwich meridian).}

\item{julian}{Julian day expressed as an integer as returned by \code{\link[=julday]{julday()}}.}

\item{merid}{an optional numeric value representing the longitude (decimal degrees) of the local time zone meridian (0 for UK time).}

\item{dst}{an optional numeric value representing the local summer time adjustment (hours, e.g. +1 for BST).}
}
\value{
a numeric value representing the solar azimuth (decimal degrees).
}
\description{
\code{solazi} is used to calculate the solar azimuth at any given location from the local time.
}
\examples{
# solar azimuth at noon on 21 June 2010, Porthleven, Cornwall, UK
jd <- julday (2010, 6, 21) # Julian day
solazi(12, 50.08, -5.31, jd)
}
