% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othertools.R
\name{suntimes}
\alias{suntimes}
\title{Calculates sunrise, sunset and daylength}
\usage{
suntimes(julian, lat, long, tz = 0, dst = 0)
}
\arguments{
\item{julian}{the Julian day as returned by \code{\link[=julday]{julday()}}.}

\item{lat}{latitude of the location for which \code{suntime} is required (decimal degrees, -ve south of equator).}

\item{long}{longitude of the location for which \code{suntime} is required (decimal degrees, -ve west of Greenwich meridian).}

\item{tz}{an optional numeric value specifying the time zones expressed as hours different from GMT (-ve to west).}

\item{dst}{an optional numeric value representing the local summer time adjustment (hours, e.g. +1 for BST).}
}
\value{
a data.frame with three components:

\code{sunrise} a vector of sunrise (hours in 24 hour clock).

\code{sunrise} a vector of sunset (hours in 24 hour clock).

\code{sunrise} a vector of daylengths (hours).
}
\description{
\code{suntimes} is used to calculate, sunrise, sunset and daylength on any given data at a specified location.
}
\details{
if the sun is above or below the horizon for 24 hours on any
days, a warning is given, and the sunrise and sunset returned are the
appoximate times at which that the sun is closest to the horizon.
}
\examples{
jd <- julday(2018, 1, 16)
suntimes(jd, 50.17, -5.12) # Cornwall, UK
suntimes(jd, 78.22, 15.64, 1) # Longyearbyen, Svalbad
suntimes(-54.94, -67.61, -3) # Puerto Williams, Cabo de Hornos, Chile
}
