% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tas}
\alias{tas}
\title{A 1km resolution dataset of daily sea-level temperature}
\format{
An array with 19 rows, 22 columns and 365 daily values
}
\source{
\url{https://eip.ceh.ac.uk/chess/}
}
\usage{
tas
}
\description{
A dataset containing daily sea-level temperature (ºC) in 2010 for the area bounded by
160000, 182000, 11000, 30000  (xmin, xmax, ymin, ymax) using the Ordance Survey GB Grid
Reference system (CRS: +init=epsg:27700).
}
\keyword{datasets}
