% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temp100}
\alias{temp100}
\title{A 100 m resolution array of hourly reference temperatures.}
\format{An array with 10 rows, 10 columns and 744 hourly values.}
\usage{
temp100
}
\description{
A 100 m resolution three-dimensional array of hourly reference temperatures
(ºC) for the area bounded by 169000, 170000, 12000, 13000  (xmin, xmax, ymin, ymax)
using the Ordance Survey GB Grid Reference system (CRS: +init=epsg:27700),
for May 2010 as output by \code{\link[=runmicro]{runmicro()}}.
}
\keyword{datasets}
