% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadtools.R
\name{updown}
\alias{updown}
\title{Internal funcion to check whether neighbouring cell is higher or lower}
\usage{
updown(dem)
}
\description{
Internal funcion to check whether neighbouring cell is higher or lower
}
