% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{veg_hgt}
\alias{veg_hgt}
\title{A 1 m resolution SpatRaster object of vegetation height.}
\format{
A SpatRaster with 1000 rows and 1000 columns.
}
\source{
\url{http://www.tellusgb.ac.uk/}
}
\usage{
veg_hgt
}
\description{
A dataset containing the vegetation height (m) for the area bounded by
169000, 170000, 12000, 13000  (xmin, xmax, ymin, ymax) using the Ordance
Survey GB Grid Reference system (CRS: +init=epsg:27700). Data were derived
from a digital terrain and digital surface model.
}
\keyword{datasets}
