% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wind1m}
\alias{wind1m}
\title{A one metre resolution matrix of windspeed}
\format{
A matrix with 1000 rows and 1000 columns.
}
\usage{
wind1m
}
\description{
A dataset containing wind speed values (\ifelse{html}{\out{m s<sup>-1</sup>}}{\eqn{m s^{-1}}})
for 2010-05-24 11:00 GMT across the area bounded by 169000, 170000, 12000, 13000
(xmin, xmax, ymin, ymax) using the Ordance Survey GB Grid Reference
system (CRS: +init=epsg:27700), as produced with \code{\link[=windcoef]{windcoef()}}
}
\keyword{datasets}
