% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othertools.R
\name{windheight}
\alias{windheight}
\title{Applies height correction to wind speed measurements}
\usage{
windheight(ui, zi, zo)
}
\arguments{
\item{ui}{numeric value(s) of measured wind speed (\ifelse{html}{\out{m s<sup>-1</sup> }}{\eqn{ m s^{-1}}}) at height \code{zi} (m).}

\item{zi}{a numeric value indicating the height (m) above the ground at which \code{ui} was measured.}

\item{zo}{a numeric value indicating the height (m) above ground level at which wind speed is desired.}
}
\value{
numeric values(s) of wind speed at height specified by \code{zo} (\ifelse{html}{\out{m s<sup>-1</sup> }}{\eqn{m s^{-1}}}).
}
\description{
\code{windheight} is used to to apply a height correction to wind speed measured at a specified height above ground level to obtain estimates of wind speed at a desired height above the ground.
}
\details{
Thus function assumes a logarithmic height profile to convert
wind speeds. It performs inaccurately when \code{uo} is lower than 0.2
and a warning is given. If \code{uo} is below ~0.08 then the logairthmic height
profile cannot be used, and \code{uo} is converted to 0.1 and a warning given.
}
\examples{
windheight(3, 10, 1) # good
windheight(3, 10, 0.15) # performs poorly. Warning given
windheight(3, 10, 0.05) # cannot calculate. ui converted and warning given

}
\seealso{
The function \code{\link[=windcoef]{windcoef()}} calculates a topographic or vegetation sheltering effect.
}
