% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadtools.R
\name{basindelin}
\alias{basindelin}
\title{Delineates hydrological basins}
\usage{
basindelin(dem)
}
\arguments{
\item{dem}{a raster object, two-dimensional array or matrix of elevations.}
}
\value{
a raster object, two-dimensional array or matrix  with individual basins numbered sequentially as integers.
}
\description{
\code{basindelin} uses digital elevation data to delineate hydrological or cold-air drainage basins.
}
\details{
If \code{dem} is a raster object, a raster onbject is returned.
This function is used to delineate cold-air drainage basins.
It iteratively identifies the lowest elevation pixel of \code{dem}, and
assigns any of the eight adjoining pixels to the same basin if higher.
The process is repeated on all assigned adjoining pixels until no further
higher pixels are found. The next lowest unassigned pixel is then
identified, a new basin identity assigned and the processes repeated until
all pixels are assigned to a basin. Relative to heuristic algorithms, it is
slow and run time increases exponentially with size of \code{dtm}. However, in
contrast to many such algorithms, all basins are correctly seperated by
boundaries >0. With large datasets, with > 160,000 pixels, calculations will
be slow and \code{\link[=basindelin_big]{basindelin_big()}} should be used instead
}
\examples{
dem <- aggregate(dtm1m, 20)
basins <- basindelin (dem)
plot(basins, main = "Basins")
}
\seealso{
\code{\link[=basindelin_big]{basindelin_big()}} for working with large datasets.
}
