% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadtools.R
\name{basindelin_big}
\alias{basindelin_big}
\title{Delineates hydrological basins for large datasets}
\usage{
basindelin_big(dem, dirout = NA, trace = TRUE)
}
\arguments{
\item{dem}{a raster object of elevations.}

\item{dirout}{an optional character vector containing a single path directory for temporarily storing tiles. Deleted after use. Tilde expansion (see \code{\link[=path.expand]{path.expand()}}) is done.}

\item{trace}{a logical value indicating whether to plot and report on progress.}
}
\value{
a raster object with individual basins numbered sequentially as integers.
}
\description{
\code{basindelin_big} is for use with large digital elevation datasets, to
delineate hydrological or cold-air drainage basins.
}
\details{
The function \code{basindelin_big} divides the large dataset into tiles and then
uses \code{\link[=basindelin]{basindelin()}} to delineate basins for each tile before mosaicing back
together and merging basins along tile edges if not seperated by a boundaryIf dirout is unspecified, then a directory basinsout istemporarily created within the working directory. If trace is TRUE (thedefault) then progress is tracked during three stages: (1) the basinsof each tile are plotted, (2) basins after mosaicing, but priorto merging are plotted and (3) on each merge iteration, the number of basinsto merge is printed and processed basin is plotted.
}
\examples{
library(raster)
basins <- basindelin_big(dtm1m)
plot(basins, main = "Basins")
}
\seealso{
\code{\link[=basindelin]{basindelin()}} for working with smaller datasets.
}
