% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demworld}
\alias{demworld}
\title{A 2.5º resolution dataset of elevations}
\format{A raster object with 72 rows and 144 columns.}
\source{
\url{http://www.cmsaf.eu/}
}
\usage{
demworld
}
\description{
A dataset global dataset elevation in metres (CRS: +init=epsg:4326)
}
\keyword{datasets}
