% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools.R
\name{hourlyNCEP}
\alias{hourlyNCEP}
\title{Interpolate NCEP data for running microclima to hourly}
\usage{
hourlyNCEP(ncepdata = NA, lat, long, tme, reanalysis2 = TRUE)
}
\arguments{
\item{ncepdata}{an optional  data frame of climate variables as returned by \code{\link[=get_NCEP]{get_NCEP()}}.}

\item{lat}{the latitude of the location for which data are required. Ignored if \code{ncepdata} provided.}

\item{long}{the longitude of the location for which data are required. Ignored if \code{ncepdata} provided.}

\item{tme}{a POSIXlt object covering the duration for which data are required. Ignored if \code{ncepdata`` provided. Hourly data are returned irrespective of the time interval of}tme\code{.Ignored if}ncepdata` provided.}

\item{reanalysis2}{Logical. Should data be obtained from the Reanalysis II dataset (default) or
from Reanalysis I (data prior to 1979). Ignored if \code{ncepdata} provided.}
}
\value{
a dataframe with the following variables:
\describe{
\item{obs_time}{POSIXlt object of times in UTC}
\item{temperature}{emperatures at 2m (ºC)}
\item{humidity}{specific humidity at 2m (Kg / Kg)}
\item{pressure}{surface pressure (Pa)}
\item{windspeed}{wind speed at 2m (metres per second}
\item{winddir}{wind direction (degrees from N)}
\item{emissivity}{emissivity of the atmosphere (0 - 1, downlong / uplong)}
\item{netlong}{Net longwave radiation (\ifelse{html}{\out{MJ m<sup>-2</sup> hr<sup>-1</sup>}}{\eqn{MJ m^-2 hr^{-1}}})}
\item{uplong}{Upward longwave radiation (\ifelse{html}{\out{MJ m<sup>-2</sup> hr<sup>-1</sup>}}{\eqn{MJ m^-2 hr^{-1}}})}
\item{downlong}{Downward longwave radiation (\ifelse{html}{\out{MJ m<sup>-2</sup> hr<sup>-1</sup>}}{\eqn{MJ m^-2 hr^{-1}}})}
\item{rad_dni}{Direct radiation normal to the solar beam (\ifelse{html}{\out{MJ m<sup>-2</sup> hr<sup>-1</sup>}}{\eqn{MJ m^-2 hr^{-1}}})}
\item{rad_dif}{Diffuse radiation (\ifelse{html}{\out{MJ m<sup>-2</sup> hr<sup>-1</sup>}}{\eqn{MJ m^-2 hr^{-1}}})}
\item{szenith}{the zenith angle (degrees)}
\item{cloudcover}{cloud cover (Percentage)}
}
}
\description{
\code{hourlyNCEP} optionally downloads the required NCEP climate and radiation forcing data
required for running microclima and interpolates 4x daily data to hourly.
}
\details{
If \code{ncepdata} is not provided, then \code{\link[=get_NCEP]{get_NCEP()}} is called and data are downloaded from NCEP Atmospheric
Model Intercomparison Project (Kanamitso et al 2002). Six-hourly data are interpolated as follows.
Pressure, humidity and the u and v wind vectors are converted to hourly using spline interpolation. Wind speeed and direction and then
calculated and adjusted to give values at 1 m using \code{\link[=windheight]{windheight()}}. The diffuse radiation
proportion is calculated using \code{\link[=difprop]{difprop()}}, and hourly obtained by adusting for the direction
of the solar beam and airmass thickness using functions \code{\link[=siflat]{siflat()}} and \code{\link[=airmasscoef]{airmasscoef()}}.
Hourly temperature are derived using \code{\link[=hourlytemp]{hourlytemp()}} and longwave radiation by splining emissivity
values.
}
\examples{
tme <- as.POSIXlt(c(0:30) * 24 * 3600, origin ="2015-01-15 00:00", tz = "UTC")
# NB takes a while to download data
hdata<- hourlyNCEP(NA, 50, -5, tme)
head(hdata)
plot(temperature ~ as.POSIXct(obs_time), data = hdata, type = "l", xlab = "Month")
plot(rad_dif ~ as.POSIXct(obs_time), data = hdata, type = "l", xlab = "Month")
}
\seealso{
\code{\link[=get_NCEP]{get_NCEP()}}
}
