% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solartools.R
\name{is_raster}
\alias{is_raster}
\title{Checks whether object is a raster and returns a matrix if yes}
\usage{
is_raster(r)
}
\arguments{
\item{r}{an R object}
}
\value{
if \code{r} is a raster, returns a matrix containing all values of \code{r}, otherwise returns \code{r}
}
\description{
\code{is_raster} is used to permit flexibility in the use of rasters, matrices or arrays in many functions.
}
\examples{
library(raster)
r <- is_raster(dtm100m)
class(dtm100m) # is a RasterLayer
class(r) # is a matrix
plot(r) # not a raster
plot(raster(r)) # converts to raster
}
