% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiationtools.R
\name{longwavetopo}
\alias{longwavetopo}
\title{calculates net longwave radiation above canopy}
\usage{
longwavetopo(h, tc, p = 101300, n, svf = 1)
}
\arguments{
\item{h}{a single numeric value, raster object, two-dimensional array or matrix of specific humidities (\ifelse{html}{\out{kg kg<sup>{-1}</sup> }}{\eqn{kg kg^{-1}}}).}

\item{tc}{a single numeric value, raster object, two-dimensional array or matrix of temperatures (ºC).}

\item{p}{an optional single numeric value, raster object, two-dimensional array or matrix of sea-level pressures (Pa).}

\item{n}{a single numeric value, raster object, two-dimensional array or matrix of fractional cloud cover values (range 0 - 1).}

\item{svf}{an optional single value, raster object, two-dimensional array or matrix of values representing the proportion of isotropic radiation received by a partially obscured surface relative to the full hemisphere, as returned by \code{\link[=skyviewtopo]{skyviewtopo()}}.}
}
\value{
a single numeric value, raster object, two-dimensional array pr matrix of values representing net longwave radiation (MJ per metre squared per hour).
}
\description{
\code{longwavetopo} is used to calculate a high-resolution dataset of the net
longwave radiation flux density emmited from the Earth, ignoring canopy
effects.
}
\details{
if \code{svf} is a raster object, a raster object is returned.
If no values for \code{p} are provided, a default value of 101300 Pa, typical of
sea-level pressure, is assumed. If single values of \code{h}, \code{tc}, \code{p} and \code{n}
are given, and \code{svf} is an array or matrix, then the entire area is assumed to
have the same values of \code{h}, \code{tc}, \code{p} and \code{n}. If no value for \code{svf} is
provided then the entire hemisphere is assumed to be in view. If single
values of \code{h}, \code{tc}, \code{p} and \code{n} are given, and no value of `svf' is
provided, a single value is returned, and it is assumed that the entire
hemisphere is in view.
}
\examples{
# =================================
# Extract data for 2010-05-24 11:00
# =================================
h <- huss[,,144]
p <- pres[,,144]
tc <- tas[,,144] + dtr[,,144]
n <-cfc[,,3444]
sv <- skyviewtopo(dtm100m)
# ===========================
# Resample to 100m resolution
# ===========================
hr <- if.raster(h, dtm1km)
tr <- if.raster(tc, dtm1km)
pr <- if.raster(p, dtm1km)
nr <- raster(n, xmn = -5.40, xmx = -5.00, ymn = 49.90, ymx = 50.15)
crs(nr) <- '+init=epsg:4326'
nr <- projectRaster(nr, crs = '+init=epsg:27700')
hr <- resample(hr, dtm100m)
tr <- resample(tr, dtm100m)
pr <- resample(pr, dtm100m)
nr <- resample(nr, dtm100m)
# =========================================
# Calculate and plot net longwave radiation
# =========================================
netlong100m <- longwavetopo(hr, tr, pr, nr, sv)
netlong100m <- mask(netlong100m, dtm100m)
plot(netlong100m, main = "Net longwave radiation")
}
\seealso{
The function \code{\link[=longwaveveg]{longwaveveg()}} returns the net longwave radiation under vegetation.
The function \code{\link[=humidityconvert]{humidityconvert()}} can be used to derive specific humidy from other meaures of humidity.
}
